Clean up refactor remnants
