+++ js/etherpad_nextcloud-embed-main.mjs
@@ -1,2 +1,2 @@
-import{o as $}from"./oc-compat-C8UpNj_V.chunk.mjs";(function(){const a=document.getElementById("etherpad-nextcloud-embed");if(!(a instanceof HTMLElement))return;const g=Number(a.getAttribute("data-file-id")||""),I=String(a.getAttribute("data-open-by-id-url")||"").trim(),j=String(a.getAttribute("data-initialize-by-id-url-template")||"").trim(),U=String(a.getAttribute("data-request-token")||"").trim(),B=String(a.getAttribute("data-trusted-origins")||"").split(/\s+/).map(e=>e.trim()).filter(Boolean),c=a.querySelector("[data-epnc-embed-loading]"),p=a.querySelector("[data-epnc-embed-error]"),F=a.querySelector("[data-epnc-embed-error-message]"),d=a.querySelector("[data-epnc-embed-iframe]"),z=String(a.getAttribute("data-l10n-external-title")||"Pad from another server").trim(),D=String(a.getAttribute("data-l10n-external-message")||"Read-only snapshot from the .pad file.").trim(),Y=String(a.getAttribute("data-l10n-external-empty")||"No synced snapshot is stored in this .pad file yet.").trim(),G=String(a.getAttribute("data-l10n-external-link")||"Open original pad").trim();let u="",P=12e4,h=null,C=!1,A=!1,b=!1,y=null,M=null,N=null,H=null;const w=()=>$(U),_=e=>{c instanceof HTMLElement&&(c.hidden=!0,c.classList.remove("epnc-embed__loading--snapshot")),d instanceof HTMLIFrameElement&&(d.hidden=!0,d.removeAttribute("src")),F instanceof HTMLElement&&(F.textContent=String(e||"Unknown error.")),p instanceof HTMLElement&&(p.hidden=!1)},J=(e,t,n)=>{if(p instanceof HTMLElement&&(p.hidden=!0),d instanceof HTMLIFrameElement&&(d.hidden=!0,d.removeAttribute("src")),!(c instanceof HTMLElement))return;c.hidden=!1,c.classList.add("epnc-embed__loading--snapshot"),c.textContent="";const r=document.createElement("div");r.className="epnc-embed__snapshot";const o=document.createElement("div");o.className="epnc-embed__snapshot-inner";const s=document.createElement("h2");s.className="epnc-embed__snapshot-title",s.textContent=z;const i=document.createElement("p");i.className="epnc-embed__snapshot-message",i.textContent=D;const l=document.createElement("a");l.className="epnc-embed__snapshot-link",l.href=e,l.target="_blank",l.rel="noopener noreferrer",l.textContent=G;const f=document.createElement("div");f.className="epnc-embed__snapshot-actions",f.appendChild(l);const m=String(n||"").trim()!=="",L=document.createElement(m?"div":"pre");L.className=m?"epnc-embed__snapshot-text epnc-embed__snapshot-text--html":"epnc-embed__snapshot-text",m?L.innerHTML=String(n):L.textContent=String(t||"").trim()!==""?String(t):Y;const T=document.createElement("div");T.className="epnc-embed__snapshot-heading",T.appendChild(s),T.appendChild(i);const x=document.createElement("div");x.className="epnc-embed__snapshot-header",x.appendChild(T),x.appendChild(f),o.appendChild(x),o.appendChild(L),r.appendChild(o),c.appendChild(r)},K=e=>{if(!(d instanceof HTMLIFrameElement)){_("Embed iframe is not available.");return}p instanceof HTMLElement&&(p.hidden=!0),c instanceof HTMLElement&&c.classList.remove("epnc-embed__loading--snapshot"),d.hidden=!0;const t=()=>{d.removeEventListener("load",t),window.setTimeout(()=>{c instanceof HTMLElement&&(c.hidden=!0),d.hidden=!1},100)};d.addEventListener("load",t,{once:!0}),d.src=e},E=()=>{y!==null&&(window.clearInterval(y),y=null)},S=(e,t)=>{v(e,t).catch(()=>{})},k=(e,t,n,r={})=>{!e||typeof e.postMessage!="function"||e.postMessage(Object.assign({type:n,fileId:g},r),t)},v=async(e,t)=>{if(!u)return{status:"disabled"};if(h)return e&&!C?(A=!0,b=b||!!t,h.catch(()=>{}).then(()=>v(!0,b))):h;C=!!e;const n=(async()=>{const l=e?u+(u.includes("?")?"&":"?")+"force=1":u,f=await fetch(l,{method:"POST",credentials:"same-origin",headers:{Accept:"application/json",requesttoken:w()},keepalive:!!t}),m=await f.json().catch(()=>({}));if(!f.ok)throw new Error(m&&m.message||"Sync request failed.");return m})();h=n;let r,o=null;try{r=await n}catch(l){o=l}finally{h===n&&(h=null),C=!1}const s=A,i=b;if(A=!1,b=!1,s)return v(!0,i);if(o instanceof Error)throw o;return r},q=()=>{!u||y!==null||(y=window.setInterval(()=>{document.visibilityState==="visible"&&S(!1,!1)},P))},Q=()=>{M||N||(M=()=>{if(document.visibilityState==="hidden"){S(!0,!0),E();return}q()},N=()=>{S(!0,!0),E()},document.addEventListener("visibilitychange",M),window.addEventListener("pagehide",N))},V=e=>!e||e==="null"?!1:e===window.location.origin?!0:B.includes(e),W=()=>{H||(H=e=>{const t=String(e.origin||"");if(!V(t))return;const n=e.data,r=typeof n=="string"?n:n&&typeof n=="object"&&typeof n.type=="string"?n.type:"";if(r){if(r==="epnc:host-visible"){q();return}if(r==="epnc:host-hidden"){S(!0,!0),E();return}if(r==="epnc:host-before-close"||r==="epnc:host-sync-now"){const o=r!=="epnc:host-sync-now",s=r==="epnc:host-before-close"?"before-close":"sync-now";k(e.source,t,"epnc:sync-flush-started",{reason:s}),v(!0,o).then(i=>{k(e.source,t,"epnc:sync-flush-finished",{reason:s,result:i&&typeof i=="object"?i:{}})}).catch(i=>{k(e.source,t,"epnc:sync-flush-failed",{reason:s,message:i instanceof Error?i.message:"Sync failed."})}),o&&E()}}},window.addEventListener("message",H))},O=async(e,t={})=>{const n=new AbortController,r=window.setTimeout(()=>n.abort(),1e4),o=Object.assign({Accept:"application/json"},t.headers||{});try{const s=await fetch(e,Object.assign({credentials:"same-origin",headers:o,signal:n.signal},t)),i=await s.json().catch(()=>({}));if(!s.ok)throw new Error(i&&i.message||"Request failed.");return i}catch(s){throw s&&typeof s=="object"&&"name"in s&&s.name==="AbortError"?new Error("Request timed out."):s}finally{window.clearTimeout(r)}},X=e=>e instanceof Error?String(e.message||"").includes("Missing YAML frontmatter"):!1,R=async()=>{const e=new URLSearchParams;e.set("fileId",String(g));const t=await O(I,{method:"POST",headers:{"Content-Type":"application/x-www-form-urlencoded;charset=UTF-8",requesttoken:w()},body:e.toString()});if(!t||typeof t.url!="string"||t.url.trim()==="")throw new Error("Pad open API did not return a valid URL.");return t},Z=async()=>{const e=j.replace("__FILE_ID__",encodeURIComponent(String(g)));await O(e,{method:"POST",headers:{requesttoken:w()}})};(async()=>{if(!Number.isFinite(g)||g<=0||I===""||j===""){_("Embed configuration is incomplete.");return}if(w()===""){_("CSRF request token is missing.");return}try{let e;try{e=await R()}catch(n){if(!X(n))throw n;await Z(),e=await R()}u=typeof e.sync_url=="string"?e.sync_url.trim():"";const t=Number(e.sync_interval_seconds??0);if(P=Number.isFinite(t)&&t>0?t*1e3:12e4,Q(),W(),q(),e.is_external===!0){J(e.url,typeof e.snapshot_text=="string"?e.snapshot_text:"",typeof e.snapshot_html=="string"?e.snapshot_html:"");return}K(e.url)}catch(e){_(e instanceof Error?e.message:"Pad open failed.")}})()})();
+import{o as $}from"./oc-compat-C8UpNj_V.chunk.mjs";import{f as R}from"./fetch-helpers-nm85FrXl.chunk.mjs";(function(){const s=document.getElementById("etherpad-nextcloud-embed");if(!(s instanceof HTMLElement))return;const g=Number(s.getAttribute("data-file-id")||""),q=String(s.getAttribute("data-open-by-id-url")||"").trim(),F=String(s.getAttribute("data-initialize-by-id-url-template")||"").trim(),U=String(s.getAttribute("data-request-token")||"").trim(),B=String(s.getAttribute("data-trusted-origins")||"").split(/\s+/).map(e=>e.trim()).filter(Boolean),i=s.querySelector("[data-epnc-embed-loading]"),p=s.querySelector("[data-epnc-embed-error]"),P=s.querySelector("[data-epnc-embed-error-message]"),a=s.querySelector("[data-epnc-embed-iframe]"),z=String(s.getAttribute("data-l10n-external-title")||"Pad from another server").trim(),D=String(s.getAttribute("data-l10n-external-message")||"Read-only snapshot from the .pad file.").trim(),Y=String(s.getAttribute("data-l10n-external-empty")||"No synced snapshot is stored in this .pad file yet.").trim(),G=String(s.getAttribute("data-l10n-external-link")||"Open original pad").trim();let u="",j=12e4,h=null,C=!1,M=!1,b=!1,y=null,A=null,N=null,H=null;const _=()=>$(U),E=e=>{i instanceof HTMLElement&&(i.hidden=!0,i.classList.remove("epnc-embed__loading--snapshot")),a instanceof HTMLIFrameElement&&(a.hidden=!0,a.removeAttribute("src")),P instanceof HTMLElement&&(P.textContent=String(e||"Unknown error.")),p instanceof HTMLElement&&(p.hidden=!1)},J=(e,t,n)=>{if(p instanceof HTMLElement&&(p.hidden=!0),a instanceof HTMLIFrameElement&&(a.hidden=!0,a.removeAttribute("src")),!(i instanceof HTMLElement))return;i.hidden=!1,i.classList.add("epnc-embed__loading--snapshot"),i.textContent="";const r=document.createElement("div");r.className="epnc-embed__snapshot";const c=document.createElement("div");c.className="epnc-embed__snapshot-inner";const l=document.createElement("h2");l.className="epnc-embed__snapshot-title",l.textContent=z;const o=document.createElement("p");o.className="epnc-embed__snapshot-message",o.textContent=D;const d=document.createElement("a");d.className="epnc-embed__snapshot-link",d.href=e,d.target="_blank",d.rel="noopener noreferrer",d.textContent=G;const f=document.createElement("div");f.className="epnc-embed__snapshot-actions",f.appendChild(d);const m=String(n||"").trim()!=="",L=document.createElement(m?"div":"pre");L.className=m?"epnc-embed__snapshot-text epnc-embed__snapshot-text--html":"epnc-embed__snapshot-text",m?L.innerHTML=String(n):L.textContent=String(t||"").trim()!==""?String(t):Y;const x=document.createElement("div");x.className="epnc-embed__snapshot-heading",x.appendChild(l),x.appendChild(o);const T=document.createElement("div");T.className="epnc-embed__snapshot-header",T.appendChild(x),T.appendChild(f),c.appendChild(T),c.appendChild(L),r.appendChild(c),i.appendChild(r)},K=e=>{if(!(a instanceof HTMLIFrameElement)){E("Embed iframe is not available.");return}p instanceof HTMLElement&&(p.hidden=!0),i instanceof HTMLElement&&i.classList.remove("epnc-embed__loading--snapshot"),a.hidden=!0;const t=()=>{a.removeEventListener("load",t),window.setTimeout(()=>{i instanceof HTMLElement&&(i.hidden=!0),a.hidden=!1},100)};a.addEventListener("load",t,{once:!0}),a.src=e},w=()=>{y!==null&&(window.clearInterval(y),y=null)},S=(e,t)=>{v(e,t).catch(()=>{})},I=(e,t,n,r={})=>{!e||typeof e.postMessage!="function"||e.postMessage(Object.assign({type:n,fileId:g},r),t)},v=async(e,t)=>{if(!u)return{status:"disabled"};if(h)return e&&!C?(M=!0,b=b||!!t,h.catch(()=>{}).then(()=>v(!0,b))):h;C=!!e;const n=(async()=>{const d=e?u+(u.includes("?")?"&":"?")+"force=1":u,f=await fetch(d,{method:"POST",credentials:"same-origin",headers:{Accept:"application/json",requesttoken:_()},keepalive:!!t}),m=await f.json().catch(()=>({}));if(!f.ok)throw new Error(m&&m.message||"Sync request failed.");return m})();h=n;let r,c=null;try{r=await n}catch(d){c=d}finally{h===n&&(h=null),C=!1}const l=M,o=b;if(M=!1,b=!1,l)return v(!0,o);if(c instanceof Error)throw c;return r},k=()=>{!u||y!==null||(y=window.setInterval(()=>{document.visibilityState==="visible"&&S(!1,!1)},j))},Q=()=>{A||N||(A=()=>{if(document.visibilityState==="hidden"){S(!0,!0),w();return}k()},N=()=>{S(!0,!0),w()},document.addEventListener("visibilitychange",A),window.addEventListener("pagehide",N))},V=e=>!e||e==="null"?!1:e===window.location.origin?!0:B.includes(e),W=()=>{H||(H=e=>{const t=String(e.origin||"");if(!V(t))return;const n=e.data,r=typeof n=="string"?n:n&&typeof n=="object"&&typeof n.type=="string"?n.type:"";if(r){if(r==="epnc:host-visible"){k();return}if(r==="epnc:host-hidden"){S(!0,!0),w();return}if(r==="epnc:host-before-close"||r==="epnc:host-sync-now"){const c=r!=="epnc:host-sync-now",l=r==="epnc:host-before-close"?"before-close":"sync-now";I(e.source,t,"epnc:sync-flush-started",{reason:l}),v(!0,c).then(o=>{I(e.source,t,"epnc:sync-flush-finished",{reason:l,result:o&&typeof o=="object"?o:{}})}).catch(o=>{I(e.source,t,"epnc:sync-flush-failed",{reason:l,message:o instanceof Error?o.message:"Sync failed."})}),c&&w()}}},window.addEventListener("message",H))},X=e=>e instanceof Error?String(e.message||"").includes("Missing YAML frontmatter"):!1,O=async()=>{const e=new URLSearchParams;e.set("fileId",String(g));const t=await R(q,{method:"POST",headers:{"Content-Type":"application/x-www-form-urlencoded;charset=UTF-8",requesttoken:_()},body:e.toString()});if(!t||typeof t.url!="string"||t.url.trim()==="")throw new Error("Pad open API did not return a valid URL.");return t},Z=async()=>{const e=F.replace("__FILE_ID__",encodeURIComponent(String(g)));await R(e,{method:"POST",headers:{requesttoken:_()}})};(async()=>{if(!Number.isFinite(g)||g<=0||q===""||F===""){E("Embed configuration is incomplete.");return}if(_()===""){E("CSRF request token is missing.");return}try{let e;try{e=await O()}catch(n){if(!X(n))throw n;await Z(),e=await O()}u=typeof e.sync_url=="string"?e.sync_url.trim():"";const t=Number(e.sync_interval_seconds??0);if(j=Number.isFinite(t)&&t>0?t*1e3:12e4,Q(),W(),k(),e.is_external===!0){J(e.url,typeof e.snapshot_text=="string"?e.snapshot_text:"",typeof e.snapshot_html=="string"?e.snapshot_html:"");return}K(e.url)}catch(e){E(e instanceof Error?e.message:"Pad open failed.")}})()})();
 //# sourceMappingURL=etherpad_nextcloud-embed-main.mjs.map
