+++ js/etherpad_nextcloud-files-main.mjs
@@ -1,2 +1,2 @@
-import{M as J,a as ve,b as K,A as m,c as L,o as le,d as Ce,e as Ee,n as xe,f as Ae}from"./oc-compat-C8UpNj_V.chunk.mjs";import{p as O,g as $,n as R,a as ce,b as _,v as U,f as H,c as Se,r as de,i as q,d as Ne,e as Pe}from"./urls-C9SLQxSc.chunk.mjs";const M=()=>!!(window.OCA&&window.OCA.Viewer&&typeof window.OCA.Viewer.open=="function"),A=()=>(window.location.pathname||"").includes("/apps/files"),D=()=>{const e=window.OCP&&window.OCP.Files&&window.OCP.Files.Router;return e&&typeof e.goToRoute=="function"?e:null},Le=e=>{const i=e instanceof Error?e.message:String(e||"");return i.includes("Redirected when going")&&i.includes("navigation guard")},P=e=>{!e||typeof e.then!="function"||Promise.resolve(e).catch(i=>{Le(i)||window.console&&typeof window.console.debug=="function"&&window.console.debug("[etherpad_nextcloud] Unexpected navigation rejection",i)})},F=e=>{const i=Number(e);return Number.isFinite(i)&&i>0?i:null},Ie=()=>{const e=O();if(e!==null)return e;const i=new URLSearchParams(window.location.search||"");return F(i.get("fileid")||"")},Me=()=>{const e=document.documentElement,i=document.body,n=(e?.className||"")+" "+(i?.className||"");if(/theme[-_]?dark|dark-theme|theme--dark/i.test(n))return!0;const r=window.matchMedia&&window.matchMedia("(prefers-color-scheme: dark)");return!!(r&&r.matches)},Z=e=>{if(!(e instanceof HTMLElement))return!1;const i=window.getComputedStyle(e);return i.display==="none"||i.visibility==="hidden"?!1:i.position==="fixed"?!0:e.offsetParent!==null},Q="etherpad_nextcloud_open",ke=e=>{if(!e)return null;const i=[e.fileId,e.fileid,e.id,e.nodeId,e.nodeid,e.file_id];for(const n of i){const r=F(n);if(r!==null)return r}if(typeof e.get=="function")for(const n of["fileid","fileId","id"])try{const r=F(e.get(n));if(r!==null)return r}catch{}return null},Te=e=>{if(!(e instanceof Element))return null;const i=[e.getAttribute("data-fileid"),e.getAttribute("data-file-id"),e.getAttribute("data-id"),e.getAttribute("data-node-id")];for(const r of i){const o=F(r);if(o!==null)return o}const n=e.matches("a[href]")?e:e.querySelector("a[href]");return n instanceof HTMLAnchorElement?ce(n.getAttribute("href")):null},Oe=e=>{if(!e||typeof e!="object")return null;const i=[e,e.fileInfoModel,e.fileInfo,e.model,e.node,e.file,e.attributes];for(const r of i){const o=ke(r);if(o!==null)return o}const n=[e.fileElement,e.el,e.element,e.$file&&e.$file[0],e.$el&&e.$el[0],e.target,e.currentTarget];for(const r of n){const o=Te(r);if(o!==null)return o}return null},Re=({openPadInNativeViewer:e})=>{window.OCA&&window.OCA.Files&&window.OCA.Files.fileActions&&(window.OCA.Files.fileActions.register(J,Q,ve(),K(m,"etherpad-icon-color"),(i,n)=>{const r=n&&n.dir||$(),o=R(r,i),a=Oe(n);e({path:o,fileId:a})},t(m,"Open in Etherpad")),window.OCA.Files.fileActions.setDefault(J,Q))},N=new Map,qe=50,Fe=300*1e3,z=async e=>{const i=String(e),n=pe(i);if(n!==null)return n;const r=L("/apps/"+m+"/api/v1/pads/resolve")+"?fileId="+encodeURIComponent(i),o=k(r,{method:"GET",headers:{Accept:"application/json"}},"Pad resolve failed.").catch(a=>{throw N.delete(i),a});return fe(i,o),o},_e=async e=>{const i="path:"+String(e),n=pe(i);if(n!==null)return n;const r=L("/apps/"+m+"/api/v1/pads/resolve")+"?file="+encodeURIComponent(e),o=k(r,{method:"GET",headers:{Accept:"application/json"}},"Pad resolve by path failed.").catch(a=>{throw N.delete(i),a});return fe(i,o),o},j=async e=>{const i=L("/apps/"+m+"/api/v1/pads/sync-status/"+encodeURIComponent(String(e)));return k(i,{method:"GET",headers:{Accept:"application/json"}},"Sync status check failed.")},He=async(e,i)=>{let n=L("/apps/"+m+"/api/v1/pads/sync/"+encodeURIComponent(String(e)));return n+="?force=1",k(n,{method:"POST",headers:{Accept:"application/json",requesttoken:le()}},"Pad sync failed.")},Ue=async(e,i)=>{const n=L("/apps/"+m+"/api/v1/pads/from-url"),r=new URLSearchParams;return r.set("file",e),r.set("padUrl",i),k(n,{method:"POST",headers:ue(),body:r.toString()},"Could not import public pad URL.")},ze=async e=>{const i=L("/apps/"+m+"/api/v1/pads"),n=new URLSearchParams;return n.set("file",e),n.set("accessMode","public"),k(i,{method:"POST",headers:ue(),body:n.toString()},"Could not create public pad.")},ue=()=>({Accept:"application/json","Content-Type":"application/x-www-form-urlencoded;charset=UTF-8",requesttoken:le()}),pe=e=>{const i=N.get(e);return i?Date.now()-i.createdAt>Fe?(N.delete(e),null):i.request:null},fe=(e,i)=>{if(!N.has(e)&&N.size>=qe){const n=N.keys().next().value;n!==void 0&&N.delete(n)}N.set(e,{createdAt:Date.now(),request:i})},k=async(e,i,n)=>{const r=await fetch(e,{...i,credentials:"same-origin"}),o=await r.json().catch(()=>({}));if(!r.ok)throw new Error(o&&o.message||n);return o},De=800,je=180,Be=120,Ve=(e,i)=>{const n=D();if(!n||!e||!Number.isFinite(e))return!1;let r=!1;return P(n.goToRoute(null,{view:"files",fileid:String(e)},{dir:de(i)})),window.setTimeout(()=>{try{const o=window.OCA.Viewer.open({path:i,onClose:$e});r=!0,P(o)}catch{}},Be),window.setTimeout(()=>{if(r)return;const o=H(e,i);window.location.assign(o)},je),!0},$e=()=>{const e=D();if(!e)return;const i={...e.query||{}};delete i.openfile,delete i.editing,P(e.goToRoute(null,e.params||{},i))},We=()=>{let e=null,i=0;return async n=>{const r=String(n.fileId??"")+"|"+String(n.path??""),o=Date.now();if(e===r&&o-i<De)return;e=r,i=o;const a=_(),y=a!==null&&a!=="",d=()=>{if(y){window.location.assign(U(a,n.path||""));return}if(A()&&n.fileId!==null&&n.fileId!==void 0&&Number.isFinite(Number(n.fileId))){const b=n.path||"/";window.location.assign(H(Number(n.fileId),b));return}const p=A()?O():null;if((n.fileId===null||n.fileId===void 0)&&p){const b=n.path||"/";window.location.assign(H(p,b));return}if(n.fileId!==null&&n.fileId!==void 0&&Number.isFinite(Number(n.fileId))){const b=n.path||"/";window.location.assign(H(Number(n.fileId),b));return}n.path&&window.location.assign(Se(n.path))};if(!M()){d();return}let u=n.path||"",s=n.fileId??null;if(!u&&n.fileId!==null&&n.fileId!==void 0)try{const p=await z(n.fileId);u=p&&typeof p.path=="string"?p.path:"",s=p&&Number.isFinite(Number(p.file_id))?Number(p.file_id):s}catch{u=""}if(!u){d();return}if((!s||!Number.isFinite(Number(s)))&&u&&!y)try{const p=await _e(u);s=p&&Number.isFinite(Number(p.file_id))?Number(p.file_id):s}catch{}if((!s||!Number.isFinite(Number(s)))&&A()){const p=O();p&&(s=p)}if(A()){if(s&&Number.isFinite(Number(s))&&Ve(Number(s),u))return;d();return}try{const p=window.OCA.Viewer.open({path:u});P(p)}catch{d()}}},Ke=120,Ge=900,Je=async(e,i={})=>{if(M()&&e.path)try{const n=Number(e.fileId);A()&&Number.isFinite(n)&&n>0&&(Ye(n,e.path,i.resolveOpenDir),await Ze()),await Qe(e.path);const r=window.OCA.Viewer.open({path:e.path,onClose:tt});P(r);return}catch{}typeof i.fallbackOpen=="function"&&await i.fallbackOpen(e)},Ze=()=>new Promise(e=>{window.setTimeout(e,Ke)}),Qe=async e=>{const i=Ce(e);i!==""&&(Ee("editor:file:created",i),await Xe())},Xe=()=>new Promise(e=>{window.setTimeout(e,Ge)}),Ye=(e,i,n)=>{const r=D();if(!r)return;const o={...r.query||{}};delete o.openfile,delete o.editing,o.dir=et(i,n),P(r.goToRoute(null,{...r.params||{},view:r.params&&r.params.view?r.params.view:"files",fileid:String(e)},o,!0))},et=(e,i)=>{if(typeof i=="function")return i(e);const n=String(e||""),r=n.lastIndexOf("/");return r>0?n.substring(0,r):"/"},tt=()=>{const e=D();if(!e)return;const i={...e.query||{}};delete i.openfile,delete i.editing,P(e.goToRoute(null,e.params||{},i))},nt=e=>{try{const i=new URL(e,window.location.origin),n=decodeURIComponent(i.pathname||"").match(/\/p\/([^/]+)$/);if(n&&n[1]){const r=n[1].replace(/[^a-zA-Z0-9._-]+/g,"-").replace(/^-+|-+$/g,"");if(r!=="")return r+".pad"}}catch{}return"Imported pad.pad"},me=e=>{const i=document.createElement("div");i.style.position="fixed",i.style.inset="0",i.style.background="rgba(0, 0, 0, 0.45)",i.style.display="flex",i.style.alignItems="center",i.style.justifyContent="center",i.style.zIndex="20000";const n=document.createElement("div");n.style.position="relative",n.style.background="var(--color-main-background, #fff)",n.style.borderRadius="10px",n.style.boxShadow="0 10px 30px rgba(0, 0, 0, 0.25)",n.style.padding="18px",n.style.width="min(460px, calc(100vw - 24px))";const r=document.createElement("button");r.type="button",r.setAttribute("aria-label",t(m,"Close")),r.textContent="×",r.style.position="absolute",r.style.top="8px",r.style.right="10px",r.style.border="none",r.style.background="transparent",r.style.fontSize="22px",r.style.cursor="pointer",r.style.lineHeight="1";const o=document.createElement("h3");return o.textContent=e,o.style.margin="0 26px 10px 0",o.style.fontSize="18px",n.appendChild(r),n.appendChild(o),i.appendChild(n),document.body.appendChild(i),{overlay:i,dialog:n,closeButton:r}},it=()=>new Promise(e=>{const{overlay:i,dialog:n,closeButton:r}=me(t(m,"Internal public pad")),o=document.createElement("label");o.textContent=t(m,"File name"),o.style.display="block",o.style.marginBottom="6px";const a=document.createElement("input");a.type="text",a.value=t(m,"Public pad")+".pad",a.style.width="100%",a.style.boxSizing="border-box",a.style.marginBottom="12px";const y=document.createElement("p");y.style.color="var(--color-error, #c62828)",y.style.margin="0 0 12px 0",y.style.minHeight="20px";const d=document.createElement("button");d.type="button",d.className="primary",d.textContent=t(m,"Create");const u=s=>{i.remove(),e(s)};r.addEventListener("click",()=>u(null)),i.addEventListener("click",s=>{s.target===i&&u(null)}),a.addEventListener("keydown",s=>{s.key==="Enter"&&(s.preventDefault(),d.click())}),d.addEventListener("click",()=>{const s=a.value.trim();if(s===""){y.textContent=t(m,"File name is required.");return}u(s)}),n.appendChild(o),n.appendChild(a),n.appendChild(y),n.appendChild(d),a.focus(),a.select()}),rt=()=>new Promise(e=>{const{overlay:i,dialog:n,closeButton:r}=me(t(m,"External public pad (URL)")),o=document.createElement("label");o.textContent=t(m,"External pad URL"),o.style.display="block",o.style.marginBottom="6px";const a=document.createElement("input");a.type="url",a.value="https://",a.placeholder="https://",a.style.width="100%",a.style.boxSizing="border-box",a.style.marginBottom="12px";const y=document.createElement("label");y.textContent=t(m,"File name"),y.style.display="block",y.style.marginBottom="6px";const d=document.createElement("input");d.type="text",d.value="Imported pad.pad",d.style.width="100%",d.style.boxSizing="border-box",d.style.marginBottom="12px";const u=document.createElement("p");u.style.color="var(--color-error, #c62828)",u.style.margin="0 0 12px 0",u.style.minHeight="20px";const s=document.createElement("button");s.type="button",s.className="primary",s.textContent=t(m,"Create");const p=w=>{i.remove(),e(w)};r.addEventListener("click",()=>p(null)),i.addEventListener("click",w=>{w.target===i&&p(null)}),a.addEventListener("blur",()=>{const w=a.value.trim();w.startsWith("http")&&(d.value=nt(w))});const b=()=>{const w=a.value.trim(),l=d.value.trim();if(w===""){u.textContent=t(m,"External pad URL is required.");return}if(l===""){u.textContent=t(m,"File name is required.");return}p({padUrl:w,name:l})};a.addEventListener("keydown",w=>{w.key==="Enter"&&(w.preventDefault(),b())}),d.addEventListener("keydown",w=>{w.key==="Enter"&&(w.preventDefault(),b())}),s.addEventListener("click",b),n.appendChild(o),n.appendChild(a),n.appendChild(y),n.appendChild(d),n.appendChild(u),n.appendChild(s),a.focus(),a.select()}),ot=()=>new Promise(e=>{const i=document.createElement("div");i.style.position="fixed",i.style.inset="0",i.style.background="rgba(0, 0, 0, 0.45)",i.style.display="flex",i.style.alignItems="center",i.style.justifyContent="center",i.style.zIndex="20000";const n=document.createElement("div");n.style.position="relative",n.style.background="var(--color-main-background, #fff)",n.style.borderRadius="10px",n.style.boxShadow="0 10px 30px rgba(0, 0, 0, 0.25)",n.style.padding="18px",n.style.width="min(420px, calc(100vw - 24px))";const r=document.createElement("h3");r.textContent=t(m,"Create public pad"),r.style.margin="0 26px 10px 0",r.style.fontSize="18px";const o=document.createElement("button");o.type="button",o.setAttribute("aria-label",t(m,"Close")),o.textContent="×",o.style.position="absolute",o.style.top="8px",o.style.right="10px",o.style.border="none",o.style.background="transparent",o.style.fontSize="22px",o.style.cursor="pointer",o.style.lineHeight="1";const a=document.createElement("p");a.textContent=t(m,"Choose how you want to create the public pad."),a.style.margin="0 0 14px 0",a.style.opacity="0.8";const y=document.createElement("div");y.style.display="grid",y.style.gap="8px";const d=document.createElement("button");d.type="button",d.className="primary",d.textContent=t(m,"Internal public pad");const u=document.createElement("button");u.type="button",u.textContent=t(m,"External public pad (URL)");const s=K(m,"etherpad-icon-color");s!==""&&(u.style.backgroundImage=`url(${s})`),u.style.backgroundRepeat="no-repeat",u.style.backgroundPosition="12px center",u.style.backgroundSize="16px 16px",u.style.paddingLeft="34px";const p=b=>{i.remove(),e(b)};o.addEventListener("click",()=>p(null)),d.addEventListener("click",()=>p("internal")),u.addEventListener("click",()=>p("external")),i.addEventListener("click",b=>{b.target===i&&p(null)}),y.appendChild(d),y.appendChild(u),n.appendChild(o),n.appendChild(r),n.appendChild(a),n.appendChild(y),i.appendChild(n),document.body.appendChild(i)}),X=e=>q(e)?e:e+".pad",at=(e,i)=>({path:e&&typeof e.file=="string"?e.file:i,fileId:e&&Number.isFinite(Number(e.file_id))?Number(e.file_id):null}),st=({openPadInNativeViewer:e})=>{const i=(o,a)=>Je(at(o,a),{fallbackOpen:e,resolveOpenDir:de}),n=async()=>{const o=await it();if(!o)return;const a=o.trim(),y=R($(),X(a));try{const d=await ze(y);await i(d,y)}catch(d){const u=d instanceof Error?d.message:t(m,"Could not create public pad.");window.alert(u)}},r=async()=>{const o=await rt();if(!o)return;const a=o.padUrl.trim(),y=o.name.trim(),d=R($(),X(y));try{const u=await Ue(d,a);await i(u,d)}catch(u){const s=u instanceof Error?u.message:t(m,"Could not import public pad URL.");window.alert(s)}};return async()=>{const o=await ot();if(o==="internal"){await n();return}o==="external"&&await r()}},B="data-epnc-sidebar-sync-panel",W="data-epnc-sidebar-sync-mount",ye=8e3,lt=6e4,we=15e3,ct=8e3,Y=()=>{const e=["[data-cy-files-sidebar]","#app-sidebar-vue .app-sidebar","#app-sidebar-vue",".app-sidebar"];for(const i of e){const n=document.querySelector(i);if(n instanceof HTMLElement)return n}return null},dt=e=>{if(!(e instanceof HTMLElement))return null;let i=e.querySelector("["+W+'="1"]');i instanceof HTMLElement||(i=document.createElement("div"),i.className="epnc-sidebar-sync-mount",i.setAttribute(W,"1"));const n=e.querySelector("[data-cy-files-sidebar-tabs], .app-sidebar-tabs");if(n&&n.parentNode instanceof HTMLElement){const r=n.parentNode;return(i.parentNode!==r||i.nextSibling!==n)&&r.insertBefore(i,n),i}return i.parentNode!==e&&e.insertBefore(i,e.firstChild),i},ee=e=>{const i=Ie();if(i!==null)return i;if(!(e instanceof HTMLElement))return null;const n=["[data-fileid]","[data-file-id]","[data-node-id]","[data-id]"];for(const o of n){const a=e.querySelector(o);if(!(a instanceof HTMLElement))continue;const y=[a.getAttribute("data-fileid"),a.getAttribute("data-file-id"),a.getAttribute("data-node-id"),a.getAttribute("data-id"),a.dataset&&(a.dataset.fileid||a.dataset.fileId)];for(const d of y){const u=F(d);if(u!==null)return u}}const r=e.querySelectorAll("a[href]");for(const o of r){if(!(o instanceof HTMLAnchorElement))continue;const a=ce(o.getAttribute("href"));if(a!==null)return a}return null},ut=e=>e==="synced"?"etherpad-icon-color":Me()?"etherpad-icon-white":"etherpad-icon-black",pt=e=>e==="synced"?t(m,"Synchronized"):e==="syncing"?t(m,"Saving..."):e==="error"?t(m,"Sync failed"):e==="unavailable"?t(m,"Sync status unavailable"):e==="loading"?t(m,"Checking sync status..."):t(m,"Sync pending"),V=e=>e==="synced"?"synced":e==="out_of_sync"?"pending":e==="unavailable"?"unavailable":"error",ft=(e,i)=>e==="synced"?Math.min(lt,i+ct):e==="pending"?ye:we,mt=()=>{let e=null,i=null,n=null,r=0;const o=()=>{n!==null&&(window.clearTimeout(n),n=null)},a=()=>{const l=document.querySelector("["+B+'="1"]');l&&l.parentNode&&l.parentNode.removeChild(l);const f=document.querySelector("["+W+'="1"]');f&&f.parentNode&&f.parentNode.removeChild(f),o()},y=(l,f)=>{if(!(l instanceof HTMLElement))return;const c=String(f||"").trim();l.setAttribute("data-public-open-url",c);const h=l.querySelector("[data-epnc-sidebar-open-public-button]");if(!(h instanceof HTMLButtonElement))return;const v=c!=="";h.style.display=v?"inline-flex":"none",h.disabled=!v},d=(l,f,c,h,v=!0)=>{if(!(l instanceof HTMLElement))return;const g=l.querySelector("[data-epnc-sidebar-sync-icon]"),E=l.querySelector("[data-epnc-sidebar-sync-label]"),C=l.querySelector("[data-epnc-sidebar-sync-button]");if(l.setAttribute("data-file-id",String(c)),l.setAttribute("data-sync-state",f),l.setAttribute("data-sync-in-flight",h?"1":"0"),E instanceof HTMLElement&&(E.textContent=pt(f)),g instanceof HTMLElement){const S=ut(f),x=K(m,S);x!==""&&(g.style.backgroundImage=`url(${x})`)}C instanceof HTMLButtonElement&&(C.disabled=h||!v||c<=0,C.setAttribute("aria-busy",h?"true":"false"))},u=(l,f)=>{if(!(f instanceof HTMLElement))return null;let c=f.querySelector("["+B+'="1"]');if(!(c instanceof HTMLElement)){c=document.createElement("section"),c.className="epnc-sidebar-sync-panel",c.setAttribute(B,"1"),c.setAttribute("data-file-id",String(l));const h=document.createElement("div");h.className="epnc-sidebar-sync-panel__header";const v=document.createElement("span");v.className="epnc-sidebar-sync-panel__icon",v.setAttribute("data-epnc-sidebar-sync-icon","1"),v.setAttribute("aria-hidden","true");const g=document.createElement("span");g.className="epnc-sidebar-sync-panel__status",g.setAttribute("data-epnc-sidebar-sync-label","1"),h.appendChild(v),h.appendChild(g);const E=document.createElement("button");E.type="button",E.className="button-vue button-vue--size-normal button-vue--vue-secondary epnc-sidebar-sync-panel__button",E.setAttribute("data-epnc-sidebar-sync-button","1"),E.textContent=t(m,"Pad in Datei speichern"),E.addEventListener("click",async S=>{S.preventDefault(),S.stopPropagation();const x=Number(c.getAttribute("data-file-id")||"");if(!(!Number.isFinite(x)||x<=0)){d(c,"syncing",x,!0,!0);try{await He(x,!0);const G=await j(x),ge=V(G&&G.status);d(c,ge,x,!1,!0)}catch{d(c,"error",x,!1,!0)}}});const C=document.createElement("button");C.type="button",C.className="button-vue button-vue--size-normal button-vue--vue-tertiary epnc-sidebar-sync-panel__button epnc-sidebar-sync-panel__button--public-open",C.setAttribute("data-epnc-sidebar-open-public-button","1"),C.textContent=t(m,"Pad in neuem Tab öffnen"),C.style.display="none",C.addEventListener("click",S=>{S.preventDefault(),S.stopPropagation();const x=String(c.getAttribute("data-public-open-url")||"").trim();x!==""&&window.open(x,"_blank","noopener,noreferrer")}),c.appendChild(h),c.appendChild(E),c.appendChild(C)}return c.parentNode!==f&&f.appendChild(c),c},s=(l,f)=>{if(o(),!(l instanceof HTMLElement)||!Number.isFinite(f)||f<=0)return;let c=ye;const h=g=>{n=window.setTimeout(()=>{v()},Math.max(1e3,g))},v=async()=>{if(!(l instanceof HTMLElement)||!document.body.contains(l)){o();return}const g=Y();if(!(g instanceof HTMLElement)||!Z(g)||!Z(l)){o();return}if(ee(g)!==f){o();return}if(document.visibilityState!=="visible"||l.getAttribute("data-sync-in-flight")==="1"){h(c);return}try{const E=await j(f),C=V(E&&E.status);d(l,C,f,!1,!0),c=ft(C,c)}catch{d(l,"error",f,!1,!0),c=we}h(c)};h(c)},p=()=>{i!==null&&(window.clearTimeout(i),i=null)},b=async()=>{const l=++r;if(!A()||_()){a();return}const f=Y();if(!(f instanceof HTMLElement)){a();return}const c=ee(f);if(!c){a();return}const h=dt(f);if(!(h instanceof HTMLElement)){a();return}let v;try{v=await z(c)}catch{a();return}if(l!==r)return;if(!v||v.is_pad!==!0){a();return}const g=u(c,h);if(!(g instanceof HTMLElement)){a();return}const E=v&&typeof v.public_open_url=="string"?v.public_open_url.trim():"";y(g,E),d(g,"loading",c,!1,!0);try{const C=await j(c);if(l!==r)return;const S=V(C&&C.status);d(g,S,c,!1,!0),s(g,c)}catch{if(l!==r)return;d(g,"error",c,!1,!0),s(g,c)}},w=(l=0)=>{p(),i=window.setTimeout(()=>{i=null,b()},Math.max(0,l))};return{installObserver:()=>{if(e!==null||!A())return;const l=new MutationObserver(f=>{for(const c of f)for(const h of c.addedNodes)if(h instanceof Element&&(h.matches("[data-cy-files-sidebar], #app-sidebar-vue, .app-sidebar")||h.querySelector("[data-cy-files-sidebar], #app-sidebar-vue, .app-sidebar"))){w(60);return}});l.observe(document.body,{childList:!0,subtree:!0}),e=l},scheduleRefresh:w}},te=m+"_public_pad",ne="icon-filetype-etherpad-nextcloud-pad",ie='<svg xmlns="http://www.w3.org/2000/svg" xml:space="preserve" fill-rule="evenodd" stroke-linejoin="round" stroke-miterlimit="2" clip-rule="evenodd" viewBox="0 0 355 355"><path fill="#5ac395" d="M317 89v177c0 28-23 51-51 51H89c-28 0-51-23-51-51V89c0-28 23-51 51-51h177c28 0 51 23 51 51"/><path fill="#4aa57f" fill-rule="nonzero" d="M240 144q-4-3-8 0-6 6 0 9a36 36 0 0 1 0 52 6 6 0 0 0 3 12q3 0 5-3c19-18 19-50 0-70"/><path fill="#479a76" fill-rule="nonzero" d="M267 122q-4-5-9 0-4 4 0 10c25 26 25 68 0 93q-4 5 0 9 5 4 9 0c30-31 30-81 0-112"/><path fill="#fff" d="M192 130v2q-1 8-10 10H76q-9-2-11-10v-2q1-10 11-10h106q9 0 10 10m24 51v2q-1 11-12 11H78q-10 0-12-11v-2q2-10 12-11h126q12 1 12 11"/><path fill="#fff" d="M216 181v2q-1 11-12 11H78q-10 0-12-11v-2q2-10 12-11h126q12 1 12 11m-57 52v1q-1 11-10 11H76q-9 0-11-11v-1q1-10 11-11h73q9 1 10 11"/><path fill="#fff" d="M159 233v1q-1 11-10 11H76q-9 0-11-11v-1q1-10 11-11h73q9 1 10 11"/></svg>',re=98,yt=120,wt=500,oe=()=>{const e=xe();return e!==null&&e>=33},T=e=>(e&&e.message?String(e.message):"").toLowerCase().includes("duplicate"),bt=e=>{if(!e||typeof e!="object")return null;if(e.hasCreatePermission===!1)return!1;if(e.hasCreatePermission===!0)return!0;if(typeof e.canCreate=="boolean")return e.canCreate;const i=Ae(),n=[e.permissions,e.permission,e.attributes&&e.attributes.permissions,typeof e.get=="function"?e.get("permissions"):null];for(const r of n){const o=Number(r);if(Number.isFinite(o))return(o&i)===i}return null},ht=(...e)=>{for(const i of e){const n=bt(i);if(n!==null)return n}return!0},gt=()=>{const e=[window.OCP&&window.OCP.Files,window.OCA&&window.OCA.Files];for(const n of e)if(!(!n||typeof n!="object")){if(typeof n.addNewFileMenuEntry=="function")return{register:n.addNewFileMenuEntry.bind(n),unregister:typeof n.removeNewFileMenuEntry=="function"?n.removeNewFileMenuEntry.bind(n):null};if(typeof n.getNewFileMenu=="function"){let r=null;try{r=n.getNewFileMenu()}catch{r=null}if(r&&typeof r.registerEntry=="function")return{register:r.registerEntry.bind(r),unregister:typeof r.unregisterEntry=="function"?r.unregisterEntry.bind(r):null}}}const i=window._nc_newfilemenu;return i&&typeof i=="object"&&typeof i.registerEntry=="function"?{register:i.registerEntry.bind(i),unregister:typeof i.unregisterEntry=="function"?i.unregisterEntry.bind(i):null}:null},vt=({isFilesAppRoute:e,onCreatePublicPad:i})=>{let n=!1,r=!1,o=!1,a=0;const y=()=>{i()},d=(...w)=>ht(...w),u=()=>({id:te,displayName:t(m,"Public pad"),...oe()?{iconSvgInline:ie}:{iconClass:ne},order:re,enabled:d,handler:y}),s=()=>({id:te,displayName:t(m,"Public pad"),templateName:t(m,"Public pad"),iconClass:ne,...oe()?{iconSvgInline:ie}:{},fileType:"file",order:re,actionHandler:y}),p=async()=>{const w=gt();if(!w)return!1;const l=u();try{return w.register(l),n=!0,!0}catch(f){return T(f)?(n=!0,!0):!1}},b=()=>{if(r)return!0;const w=s(),l=()=>{const f=[window.OCA&&window.OCA.Files&&window.OCA.Files.NewFileMenu,window.OCA&&window.OCA.Files&&window.OCA.Files.newFileMenu];for(const c of f)if(!(!c||typeof c!="object")){if(typeof c.addMenuEntry=="function")try{return c.addMenuEntry(w),!0}catch(h){if(T(h))return!0}if(typeof c.registerEntry=="function")try{return c.registerEntry(u()),!0}catch(h){if(T(h))return!0}}return!1};try{return l()?(r=!0,!0):(!o&&window.OC&&window.OC.Plugins&&typeof window.OC.Plugins.register=="function"&&(window.OC.Plugins.register("OCA.Files.NewFileMenu",{attach(f){if(!(!f||typeof f.addMenuEntry!="function"))try{f.addMenuEntry(w),r=!0}catch(c){T(c)&&(r=!0)}}}),o=!0),r)}catch(f){return T(f)?(r=!0,!0):!1}};return()=>{if(!e()||n||r)return;const w=++a,l=async f=>{w!==a||n||r||await p()||b()||f<yt&&window.setTimeout(()=>{l(f+1)},wt)};l(0)}},Ct=(e,i)=>{if(!(e instanceof HTMLAnchorElement)||!i)return null;const n=e.getAttribute("href")||"",r=Ne(n);if(r&&r.token===i&&q(r.path))return r.path;const o=Pe(n);return q(o)?o:null},Et=({openPadInNativeViewer:e})=>{if(window.OCA&&window.OCA.EtherpadNextcloudClickInterceptorRegistered===!0)return;window.OCA&&(window.OCA.EtherpadNextcloudClickInterceptorRegistered=!0);const i=async n=>{if(!n||n.defaultPrevented)return;const r=_();if(!r||n.type==="click"&&(n.button!==0||n.metaKey||n.ctrlKey||n.shiftKey||n.altKey)||!(n.target instanceof Element))return;const o=n.target.closest("a[href]");if(!(o instanceof HTMLAnchorElement))return;const a=Ct(o,r);if(a){if(n.preventDefault(),n.stopPropagation(),typeof n.stopImmediatePropagation=="function"&&n.stopImmediatePropagation(),M()){await e({path:a,fileId:null});return}window.location.assign(U(r,a))}};document.addEventListener("click",i,!0)},I="data-epnc-single-file-share";let ae=0;const be=e=>{if(!e)return null;const i='a[href*="/s/'+e+'/download"],a[href*="/index.php/s/'+e+'/download"]',n=document.querySelector(i);return n instanceof HTMLAnchorElement?n:null},xt=()=>{const e=document.getElementById("initial-state-files_sharing-view");if(!(e instanceof HTMLInputElement))return"";const i=String(e.value||"").trim();if(i==="")return"";try{const n=window.atob(i);if(n.startsWith('"')||n.startsWith("{")||n.startsWith("[")){const r=JSON.parse(n);return typeof r=="string"?r:""}return n}catch{return""}},he=e=>{if(!e)return!1;const i=be(e);if(!i)return!1;let n;try{n=new URL(i.getAttribute("href")||"",window.location.origin)}catch{return!1}const r=(n.pathname||"").match(/(?:\/index\.php)?\/s\/([^/]+)\/download\/?$/);return!r||r[1]!==e?!1:!n.searchParams.has("files")},At=()=>{const e=document.body;if(!(e instanceof HTMLElement))return{retry:!0};const i=_();if(!i)return e.removeAttribute(I),{retry:!1};const n=xt();return n==="public-file-share"?(e.setAttribute(I,"1"),{retry:!1}):n==="public-share"?(e.removeAttribute(I),{retry:!1}):be(i)?he(i)?(e.setAttribute(I,"1"),{retry:!1}):(e.removeAttribute(I),{retry:!1}):(e.removeAttribute(I),{retry:!0})},St=()=>{const e=++ae,i=n=>{e===ae&&At().retry&&n<30&&window.setTimeout(()=>i(n+1),180)};i(0)},se="EtherpadNextcloudRouteWatchInstalled",Nt=({ensurePublicPadMenuRegistration:e,openPadInNativeViewer:i,schedulePublicSingleShareUiStateRefresh:n,scheduleSidebarSyncPanelRefresh:r})=>{let o="";const a=async()=>{if(M())return;const s=O();if(!s)return;const p=new URLSearchParams(window.location.search||"");if(!(p.get("epNoRedirect")==="1"||p.get("details")==="1"||p.get("opendetails")==="true")&&!(window.location.pathname||"").includes("/apps/"+m+"/"))try{const b=await z(s);b&&b.is_pad&&b.viewer_url&&window.location.assign(b.viewer_url)}catch{}},y=async()=>{if(M()||!A())return;const s=O();if(!s)return;const p=new URLSearchParams(window.location.search||"");if(!(p.get("openfile")==="true"||p.get("opendetails")==="true"||p.get("details")==="1"))try{const b=await z(s);if(!b||!b.is_pad)return;const w=p.get("dir")||"/",l=L("/apps/files/files")+"?dir="+encodeURIComponent(w);window.location.href!==l&&window.location.assign(l)}catch{}},d=()=>{const s=_();if(!s)return;const p=new URLSearchParams(window.location.search||""),b=p.get("files")||"";if(M()){if(!q(b))return;const f=p.get("path")||"/",c=R(f,b);i({path:c,fileId:null});return}if(!q(b)){if((window.location.pathname||"").includes("/apps/"+m+"/public/"))return;he(s)&&window.location.assign(U(s,""));return}const w=p.get("path")||"/",l=R(w,b);window.location.assign(U(s,l))},u=()=>{const s=(window.location.pathname||"")+"?"+(window.location.search||"");s!==o&&(o=s,A()&&e(),n(),y(),a(),d(),r(0))};return{evaluateCurrentRoute:u,installRouteWatchers:()=>{if(window.OCA&&window.OCA[se]===!0)return;window.OCA&&(window.OCA[se]=!0);const s=()=>window.setTimeout(u,0);window.addEventListener("popstate",s),window.addEventListener("hashchange",s);const p=b=>{const w=window.history&&window.history[b];typeof w=="function"&&(window.history[b]=function(...l){const f=w.apply(this,l);return s(),f})};p("pushState"),p("replaceState")}}};(function(){let e=!1;const i=mt(),n=We(),r=st({openPadInNativeViewer:n}),o=vt({isFilesAppRoute:A,onCreatePublicPad:r}),a=Nt({ensurePublicPadMenuRegistration:o,openPadInNativeViewer:n,schedulePublicSingleShareUiStateRefresh:St,scheduleSidebarSyncPanelRefresh:i.scheduleRefresh}),y=()=>{e||(e=!0,a.installRouteWatchers(),a.evaluateCurrentRoute(),Re({openPadInNativeViewer:n}),Et({openPadInNativeViewer:n}),i.installObserver(),i.scheduleRefresh(200))};document.readyState==="loading"?document.addEventListener("DOMContentLoaded",y,{once:!0}):y()})();
+import{M as J,a as ve,b as K,A as m,c as L,o as le,d as Ce,e as Ee,n as xe,f as Ae}from"./oc-compat-C8UpNj_V.chunk.mjs";import{g as $,n as O,p as ce,a as _,b as q,v as U,f as H,c as Se,r as de,i as R,d as Ne,e as Pe}from"./urls-BqezDGeR.chunk.mjs";const M=()=>!!(window.OCA&&window.OCA.Viewer&&typeof window.OCA.Viewer.open=="function"),A=()=>(window.location.pathname||"").includes("/apps/files"),D=()=>{const e=window.OCP&&window.OCP.Files&&window.OCP.Files.Router;return e&&typeof e.goToRoute=="function"?e:null},Le=e=>{const i=e instanceof Error?e.message:String(e||"");return i.includes("Redirected when going")&&i.includes("navigation guard")},P=e=>{!e||typeof e.then!="function"||Promise.resolve(e).catch(i=>{Le(i)||window.console&&typeof window.console.debug=="function"&&window.console.debug("[etherpad_nextcloud] Unexpected navigation rejection",i)})},F=e=>{const i=Number(e);return Number.isFinite(i)&&i>0?i:null},Ie=()=>{const e=document.documentElement,i=document.body,n=(e?.className||"")+" "+(i?.className||"");if(/theme[-_]?dark|dark-theme|theme--dark/i.test(n))return!0;const r=window.matchMedia&&window.matchMedia("(prefers-color-scheme: dark)");return!!(r&&r.matches)},Z=e=>{if(!(e instanceof HTMLElement))return!1;const i=window.getComputedStyle(e);return i.display==="none"||i.visibility==="hidden"?!1:i.position==="fixed"?!0:e.offsetParent!==null},Q="etherpad_nextcloud_open",Me=e=>{if(!e)return null;const i=[e.fileId,e.fileid,e.id,e.nodeId,e.nodeid,e.file_id];for(const n of i){const r=F(n);if(r!==null)return r}if(typeof e.get=="function")for(const n of["fileid","fileId","id"])try{const r=F(e.get(n));if(r!==null)return r}catch{}return null},ke=e=>{if(!(e instanceof Element))return null;const i=[e.getAttribute("data-fileid"),e.getAttribute("data-file-id"),e.getAttribute("data-id"),e.getAttribute("data-node-id")];for(const r of i){const o=F(r);if(o!==null)return o}const n=e.matches("a[href]")?e:e.querySelector("a[href]");return n instanceof HTMLAnchorElement?ce(n.getAttribute("href")):null},Te=e=>{if(!e||typeof e!="object")return null;const i=[e,e.fileInfoModel,e.fileInfo,e.model,e.node,e.file,e.attributes];for(const r of i){const o=Me(r);if(o!==null)return o}const n=[e.fileElement,e.el,e.element,e.$file&&e.$file[0],e.$el&&e.$el[0],e.target,e.currentTarget];for(const r of n){const o=ke(r);if(o!==null)return o}return null},Oe=({openPadInNativeViewer:e})=>{window.OCA&&window.OCA.Files&&window.OCA.Files.fileActions&&(window.OCA.Files.fileActions.register(J,Q,ve(),K(m,"etherpad-icon-color"),(i,n)=>{const r=n&&n.dir||$(),o=O(r,i),a=Te(n);e({path:o,fileId:a})},t(m,"Open in Etherpad")),window.OCA.Files.fileActions.setDefault(J,Q))},N=new Map,qe=50,Re=300*1e3,z=async e=>{const i=String(e),n=pe(i);if(n!==null)return n;const r=L("/apps/"+m+"/api/v1/pads/resolve")+"?fileId="+encodeURIComponent(i),o=k(r,{method:"GET",headers:{Accept:"application/json"}},"Pad resolve failed.").catch(a=>{throw N.delete(i),a});return fe(i,o),o},Fe=async e=>{const i="path:"+String(e),n=pe(i);if(n!==null)return n;const r=L("/apps/"+m+"/api/v1/pads/resolve")+"?file="+encodeURIComponent(e),o=k(r,{method:"GET",headers:{Accept:"application/json"}},"Pad resolve by path failed.").catch(a=>{throw N.delete(i),a});return fe(i,o),o},j=async e=>{const i=L("/apps/"+m+"/api/v1/pads/sync-status/"+encodeURIComponent(String(e)));return k(i,{method:"GET",headers:{Accept:"application/json"}},"Sync status check failed.")},_e=async(e,i)=>{let n=L("/apps/"+m+"/api/v1/pads/sync/"+encodeURIComponent(String(e)));return n+="?force=1",k(n,{method:"POST",headers:{Accept:"application/json",requesttoken:le()}},"Pad sync failed.")},He=async(e,i)=>{const n=L("/apps/"+m+"/api/v1/pads/from-url"),r=new URLSearchParams;return r.set("file",e),r.set("padUrl",i),k(n,{method:"POST",headers:ue(),body:r.toString()},"Could not import public pad URL.")},Ue=async e=>{const i=L("/apps/"+m+"/api/v1/pads"),n=new URLSearchParams;return n.set("file",e),n.set("accessMode","public"),k(i,{method:"POST",headers:ue(),body:n.toString()},"Could not create public pad.")},ue=()=>({Accept:"application/json","Content-Type":"application/x-www-form-urlencoded;charset=UTF-8",requesttoken:le()}),pe=e=>{const i=N.get(e);return i?Date.now()-i.createdAt>Re?(N.delete(e),null):i.request:null},fe=(e,i)=>{if(!N.has(e)&&N.size>=qe){const n=N.keys().next().value;n!==void 0&&N.delete(n)}N.set(e,{createdAt:Date.now(),request:i})},k=async(e,i,n)=>{const r=await fetch(e,{...i,credentials:"same-origin"}),o=await r.json().catch(()=>({}));if(!r.ok)throw new Error(o&&o.message||n);return o},ze=800,De=180,je=120,Be=(e,i)=>{const n=D();if(!n||!e||!Number.isFinite(e))return!1;let r=!1;return P(n.goToRoute(null,{view:"files",fileid:String(e)},{dir:de(i)})),window.setTimeout(()=>{try{const o=window.OCA.Viewer.open({path:i,onClose:Ve});r=!0,P(o)}catch{}},je),window.setTimeout(()=>{if(r)return;const o=H(e,i);window.location.assign(o)},De),!0},Ve=()=>{const e=D();if(!e)return;const i={...e.query||{}};delete i.openfile,delete i.editing,P(e.goToRoute(null,e.params||{},i))},$e=()=>{let e=null,i=0;return async n=>{const r=String(n.fileId??"")+"|"+String(n.path??""),o=Date.now();if(e===r&&o-i<ze)return;e=r,i=o;const a=_(),y=a!==null&&a!=="",s=()=>{if(y){window.location.assign(U(a,n.path||""));return}if(A()&&n.fileId!==null&&n.fileId!==void 0&&Number.isFinite(Number(n.fileId))){const b=n.path||"/";window.location.assign(H(Number(n.fileId),b));return}const u=A()?q():null;if((n.fileId===null||n.fileId===void 0)&&u){const b=n.path||"/";window.location.assign(H(u,b));return}if(n.fileId!==null&&n.fileId!==void 0&&Number.isFinite(Number(n.fileId))){const b=n.path||"/";window.location.assign(H(Number(n.fileId),b));return}n.path&&window.location.assign(Se(n.path))};if(!M()){s();return}let p=n.path||"",l=n.fileId??null;if(!p&&n.fileId!==null&&n.fileId!==void 0)try{const u=await z(n.fileId);p=u&&typeof u.path=="string"?u.path:"",l=u&&Number.isFinite(Number(u.file_id))?Number(u.file_id):l}catch{p=""}if(!p){s();return}if((!l||!Number.isFinite(Number(l)))&&p&&!y)try{const u=await Fe(p);l=u&&Number.isFinite(Number(u.file_id))?Number(u.file_id):l}catch{}if((!l||!Number.isFinite(Number(l)))&&A()){const u=q();u&&(l=u)}if(A()){if(l&&Number.isFinite(Number(l))&&Be(Number(l),p))return;s();return}try{const u=window.OCA.Viewer.open({path:p});P(u)}catch{s()}}},We=120,Ke=900,Ge=async(e,i={})=>{if(M()&&e.path)try{const n=Number(e.fileId);A()&&Number.isFinite(n)&&n>0&&(Xe(n,e.path,i.resolveOpenDir),await Je()),await Ze(e.path);const r=window.OCA.Viewer.open({path:e.path,onClose:et});P(r);return}catch{}typeof i.fallbackOpen=="function"&&await i.fallbackOpen(e)},Je=()=>new Promise(e=>{window.setTimeout(e,We)}),Ze=async e=>{const i=Ce(e);i!==""&&(Ee("editor:file:created",i),await Qe())},Qe=()=>new Promise(e=>{window.setTimeout(e,Ke)}),Xe=(e,i,n)=>{const r=D();if(!r)return;const o={...r.query||{}};delete o.openfile,delete o.editing,o.dir=Ye(i,n),P(r.goToRoute(null,{...r.params||{},view:r.params&&r.params.view?r.params.view:"files",fileid:String(e)},o,!0))},Ye=(e,i)=>{if(typeof i=="function")return i(e);const n=String(e||""),r=n.lastIndexOf("/");return r>0?n.substring(0,r):"/"},et=()=>{const e=D();if(!e)return;const i={...e.query||{}};delete i.openfile,delete i.editing,P(e.goToRoute(null,e.params||{},i))},tt=e=>{try{const i=new URL(e,window.location.origin),n=decodeURIComponent(i.pathname||"").match(/\/p\/([^/]+)$/);if(n&&n[1]){const r=n[1].replace(/[^a-zA-Z0-9._-]+/g,"-").replace(/^-+|-+$/g,"");if(r!=="")return r+".pad"}}catch{}return"Imported pad.pad"},me=e=>{const i=document.createElement("div");i.style.position="fixed",i.style.inset="0",i.style.background="rgba(0, 0, 0, 0.45)",i.style.display="flex",i.style.alignItems="center",i.style.justifyContent="center",i.style.zIndex="20000";const n=document.createElement("div");n.style.position="relative",n.style.background="var(--color-main-background, #fff)",n.style.borderRadius="10px",n.style.boxShadow="0 10px 30px rgba(0, 0, 0, 0.25)",n.style.padding="18px",n.style.width="min(460px, calc(100vw - 24px))";const r=document.createElement("button");r.type="button",r.setAttribute("aria-label",t(m,"Close")),r.textContent="×",r.style.position="absolute",r.style.top="8px",r.style.right="10px",r.style.border="none",r.style.background="transparent",r.style.fontSize="22px",r.style.cursor="pointer",r.style.lineHeight="1";const o=document.createElement("h3");return o.textContent=e,o.style.margin="0 26px 10px 0",o.style.fontSize="18px",n.appendChild(r),n.appendChild(o),i.appendChild(n),document.body.appendChild(i),{overlay:i,dialog:n,closeButton:r}},nt=()=>new Promise(e=>{const{overlay:i,dialog:n,closeButton:r}=me(t(m,"Internal public pad")),o=document.createElement("label");o.textContent=t(m,"File name"),o.style.display="block",o.style.marginBottom="6px";const a=document.createElement("input");a.type="text",a.value=t(m,"Public pad")+".pad",a.style.width="100%",a.style.boxSizing="border-box",a.style.marginBottom="12px";const y=document.createElement("p");y.style.color="var(--color-error, #c62828)",y.style.margin="0 0 12px 0",y.style.minHeight="20px";const s=document.createElement("button");s.type="button",s.className="primary",s.textContent=t(m,"Create");const p=l=>{i.remove(),e(l)};r.addEventListener("click",()=>p(null)),i.addEventListener("click",l=>{l.target===i&&p(null)}),a.addEventListener("keydown",l=>{l.key==="Enter"&&(l.preventDefault(),s.click())}),s.addEventListener("click",()=>{const l=a.value.trim();if(l===""){y.textContent=t(m,"File name is required.");return}p(l)}),n.appendChild(o),n.appendChild(a),n.appendChild(y),n.appendChild(s),a.focus(),a.select()}),it=()=>new Promise(e=>{const{overlay:i,dialog:n,closeButton:r}=me(t(m,"External public pad (URL)")),o=document.createElement("label");o.textContent=t(m,"External pad URL"),o.style.display="block",o.style.marginBottom="6px";const a=document.createElement("input");a.type="url",a.value="https://",a.placeholder="https://",a.style.width="100%",a.style.boxSizing="border-box",a.style.marginBottom="12px";const y=document.createElement("label");y.textContent=t(m,"File name"),y.style.display="block",y.style.marginBottom="6px";const s=document.createElement("input");s.type="text",s.value="Imported pad.pad",s.style.width="100%",s.style.boxSizing="border-box",s.style.marginBottom="12px";const p=document.createElement("p");p.style.color="var(--color-error, #c62828)",p.style.margin="0 0 12px 0",p.style.minHeight="20px";const l=document.createElement("button");l.type="button",l.className="primary",l.textContent=t(m,"Create");const u=w=>{i.remove(),e(w)};r.addEventListener("click",()=>u(null)),i.addEventListener("click",w=>{w.target===i&&u(null)}),a.addEventListener("blur",()=>{const w=a.value.trim();w.startsWith("http")&&(s.value=tt(w))});const b=()=>{const w=a.value.trim(),c=s.value.trim();if(w===""){p.textContent=t(m,"External pad URL is required.");return}if(c===""){p.textContent=t(m,"File name is required.");return}u({padUrl:w,name:c})};a.addEventListener("keydown",w=>{w.key==="Enter"&&(w.preventDefault(),b())}),s.addEventListener("keydown",w=>{w.key==="Enter"&&(w.preventDefault(),b())}),l.addEventListener("click",b),n.appendChild(o),n.appendChild(a),n.appendChild(y),n.appendChild(s),n.appendChild(p),n.appendChild(l),a.focus(),a.select()}),rt=()=>new Promise(e=>{const i=document.createElement("div");i.style.position="fixed",i.style.inset="0",i.style.background="rgba(0, 0, 0, 0.45)",i.style.display="flex",i.style.alignItems="center",i.style.justifyContent="center",i.style.zIndex="20000";const n=document.createElement("div");n.style.position="relative",n.style.background="var(--color-main-background, #fff)",n.style.borderRadius="10px",n.style.boxShadow="0 10px 30px rgba(0, 0, 0, 0.25)",n.style.padding="18px",n.style.width="min(420px, calc(100vw - 24px))";const r=document.createElement("h3");r.textContent=t(m,"Create public pad"),r.style.margin="0 26px 10px 0",r.style.fontSize="18px";const o=document.createElement("button");o.type="button",o.setAttribute("aria-label",t(m,"Close")),o.textContent="×",o.style.position="absolute",o.style.top="8px",o.style.right="10px",o.style.border="none",o.style.background="transparent",o.style.fontSize="22px",o.style.cursor="pointer",o.style.lineHeight="1";const a=document.createElement("p");a.textContent=t(m,"Choose how you want to create the public pad."),a.style.margin="0 0 14px 0",a.style.opacity="0.8";const y=document.createElement("div");y.style.display="grid",y.style.gap="8px";const s=document.createElement("button");s.type="button",s.className="primary",s.textContent=t(m,"Internal public pad");const p=document.createElement("button");p.type="button",p.textContent=t(m,"External public pad (URL)");const l=K(m,"etherpad-icon-color");l!==""&&(p.style.backgroundImage=`url(${l})`),p.style.backgroundRepeat="no-repeat",p.style.backgroundPosition="12px center",p.style.backgroundSize="16px 16px",p.style.paddingLeft="34px";const u=b=>{i.remove(),e(b)};o.addEventListener("click",()=>u(null)),s.addEventListener("click",()=>u("internal")),p.addEventListener("click",()=>u("external")),i.addEventListener("click",b=>{b.target===i&&u(null)}),y.appendChild(s),y.appendChild(p),n.appendChild(o),n.appendChild(r),n.appendChild(a),n.appendChild(y),i.appendChild(n),document.body.appendChild(i)}),X=e=>R(e)?e:e+".pad",ot=(e,i)=>({path:e&&typeof e.file=="string"?e.file:i,fileId:e&&Number.isFinite(Number(e.file_id))?Number(e.file_id):null}),at=({openPadInNativeViewer:e})=>{const i=(o,a)=>Ge(ot(o,a),{fallbackOpen:e,resolveOpenDir:de}),n=async()=>{const o=await nt();if(!o)return;const a=o.trim(),y=O($(),X(a));try{const s=await Ue(y);await i(s,y)}catch(s){const p=s instanceof Error?s.message:t(m,"Could not create public pad.");window.alert(p)}},r=async()=>{const o=await it();if(!o)return;const a=o.padUrl.trim(),y=o.name.trim(),s=O($(),X(y));try{const p=await He(s,a);await i(p,s)}catch(p){const l=p instanceof Error?p.message:t(m,"Could not import public pad URL.");window.alert(l)}};return async()=>{const o=await rt();if(o==="internal"){await n();return}o==="external"&&await r()}},B="data-epnc-sidebar-sync-panel",W="data-epnc-sidebar-sync-mount",ye=8e3,st=6e4,we=15e3,lt=8e3,Y=()=>{const e=["[data-cy-files-sidebar]","#app-sidebar-vue .app-sidebar","#app-sidebar-vue",".app-sidebar"];for(const i of e){const n=document.querySelector(i);if(n instanceof HTMLElement)return n}return null},ct=e=>{if(!(e instanceof HTMLElement))return null;let i=e.querySelector("["+W+'="1"]');i instanceof HTMLElement||(i=document.createElement("div"),i.className="epnc-sidebar-sync-mount",i.setAttribute(W,"1"));const n=e.querySelector("[data-cy-files-sidebar-tabs], .app-sidebar-tabs");if(n&&n.parentNode instanceof HTMLElement){const r=n.parentNode;return(i.parentNode!==r||i.nextSibling!==n)&&r.insertBefore(i,n),i}return i.parentNode!==e&&e.insertBefore(i,e.firstChild),i},ee=e=>{const i=q();if(i!==null)return i;const n=new URLSearchParams(window.location.search||""),r=F(n.get("fileid")||"");if(r!==null)return r;if(!(e instanceof HTMLElement))return null;const o=["[data-fileid]","[data-file-id]","[data-node-id]","[data-id]"];for(const y of o){const s=e.querySelector(y);if(!(s instanceof HTMLElement))continue;const p=[s.getAttribute("data-fileid"),s.getAttribute("data-file-id"),s.getAttribute("data-node-id"),s.getAttribute("data-id"),s.dataset&&(s.dataset.fileid||s.dataset.fileId)];for(const l of p){const u=F(l);if(u!==null)return u}}const a=e.querySelectorAll("a[href]");for(const y of a){if(!(y instanceof HTMLAnchorElement))continue;const s=ce(y.getAttribute("href"));if(s!==null)return s}return null},dt=e=>e==="synced"?"etherpad-icon-color":Ie()?"etherpad-icon-white":"etherpad-icon-black",ut=e=>e==="synced"?t(m,"Synchronized"):e==="syncing"?t(m,"Saving..."):e==="error"?t(m,"Sync failed"):e==="unavailable"?t(m,"Sync status unavailable"):e==="loading"?t(m,"Checking sync status..."):t(m,"Sync pending"),V=e=>e==="synced"?"synced":e==="out_of_sync"?"pending":e==="unavailable"?"unavailable":"error",pt=(e,i)=>e==="synced"?Math.min(st,i+lt):e==="pending"?ye:we,ft=()=>{let e=null,i=null,n=null,r=0;const o=()=>{n!==null&&(window.clearTimeout(n),n=null)},a=()=>{const c=document.querySelector("["+B+'="1"]');c&&c.parentNode&&c.parentNode.removeChild(c);const f=document.querySelector("["+W+'="1"]');f&&f.parentNode&&f.parentNode.removeChild(f),o()},y=(c,f)=>{if(!(c instanceof HTMLElement))return;const d=String(f||"").trim();c.setAttribute("data-public-open-url",d);const h=c.querySelector("[data-epnc-sidebar-open-public-button]");if(!(h instanceof HTMLButtonElement))return;const v=d!=="";h.style.display=v?"inline-flex":"none",h.disabled=!v},s=(c,f,d,h,v=!0)=>{if(!(c instanceof HTMLElement))return;const g=c.querySelector("[data-epnc-sidebar-sync-icon]"),E=c.querySelector("[data-epnc-sidebar-sync-label]"),C=c.querySelector("[data-epnc-sidebar-sync-button]");if(c.setAttribute("data-file-id",String(d)),c.setAttribute("data-sync-state",f),c.setAttribute("data-sync-in-flight",h?"1":"0"),E instanceof HTMLElement&&(E.textContent=ut(f)),g instanceof HTMLElement){const S=dt(f),x=K(m,S);x!==""&&(g.style.backgroundImage=`url(${x})`)}C instanceof HTMLButtonElement&&(C.disabled=h||!v||d<=0,C.setAttribute("aria-busy",h?"true":"false"))},p=(c,f)=>{if(!(f instanceof HTMLElement))return null;let d=f.querySelector("["+B+'="1"]');if(!(d instanceof HTMLElement)){d=document.createElement("section"),d.className="epnc-sidebar-sync-panel",d.setAttribute(B,"1"),d.setAttribute("data-file-id",String(c));const h=document.createElement("div");h.className="epnc-sidebar-sync-panel__header";const v=document.createElement("span");v.className="epnc-sidebar-sync-panel__icon",v.setAttribute("data-epnc-sidebar-sync-icon","1"),v.setAttribute("aria-hidden","true");const g=document.createElement("span");g.className="epnc-sidebar-sync-panel__status",g.setAttribute("data-epnc-sidebar-sync-label","1"),h.appendChild(v),h.appendChild(g);const E=document.createElement("button");E.type="button",E.className="button-vue button-vue--size-normal button-vue--vue-secondary epnc-sidebar-sync-panel__button",E.setAttribute("data-epnc-sidebar-sync-button","1"),E.textContent=t(m,"Pad in Datei speichern"),E.addEventListener("click",async S=>{S.preventDefault(),S.stopPropagation();const x=Number(d.getAttribute("data-file-id")||"");if(!(!Number.isFinite(x)||x<=0)){s(d,"syncing",x,!0,!0);try{await _e(x,!0);const G=await j(x),ge=V(G&&G.status);s(d,ge,x,!1,!0)}catch{s(d,"error",x,!1,!0)}}});const C=document.createElement("button");C.type="button",C.className="button-vue button-vue--size-normal button-vue--vue-tertiary epnc-sidebar-sync-panel__button epnc-sidebar-sync-panel__button--public-open",C.setAttribute("data-epnc-sidebar-open-public-button","1"),C.textContent=t(m,"Pad in neuem Tab öffnen"),C.style.display="none",C.addEventListener("click",S=>{S.preventDefault(),S.stopPropagation();const x=String(d.getAttribute("data-public-open-url")||"").trim();x!==""&&window.open(x,"_blank","noopener,noreferrer")}),d.appendChild(h),d.appendChild(E),d.appendChild(C)}return d.parentNode!==f&&f.appendChild(d),d},l=(c,f)=>{if(o(),!(c instanceof HTMLElement)||!Number.isFinite(f)||f<=0)return;let d=ye;const h=g=>{n=window.setTimeout(()=>{v()},Math.max(1e3,g))},v=async()=>{if(!(c instanceof HTMLElement)||!document.body.contains(c)){o();return}const g=Y();if(!(g instanceof HTMLElement)||!Z(g)||!Z(c)){o();return}if(ee(g)!==f){o();return}if(document.visibilityState!=="visible"||c.getAttribute("data-sync-in-flight")==="1"){h(d);return}try{const E=await j(f),C=V(E&&E.status);s(c,C,f,!1,!0),d=pt(C,d)}catch{s(c,"error",f,!1,!0),d=we}h(d)};h(d)},u=()=>{i!==null&&(window.clearTimeout(i),i=null)},b=async()=>{const c=++r;if(!A()||_()){a();return}const f=Y();if(!(f instanceof HTMLElement)){a();return}const d=ee(f);if(!d){a();return}const h=ct(f);if(!(h instanceof HTMLElement)){a();return}let v;try{v=await z(d)}catch{a();return}if(c!==r)return;if(!v||v.is_pad!==!0){a();return}const g=p(d,h);if(!(g instanceof HTMLElement)){a();return}const E=v&&typeof v.public_open_url=="string"?v.public_open_url.trim():"";y(g,E),s(g,"loading",d,!1,!0);try{const C=await j(d);if(c!==r)return;const S=V(C&&C.status);s(g,S,d,!1,!0),l(g,d)}catch{if(c!==r)return;s(g,"error",d,!1,!0),l(g,d)}},w=(c=0)=>{u(),i=window.setTimeout(()=>{i=null,b()},Math.max(0,c))};return{installObserver:()=>{if(e!==null||!A())return;const c=new MutationObserver(f=>{for(const d of f)for(const h of d.addedNodes)if(h instanceof Element&&(h.matches("[data-cy-files-sidebar], #app-sidebar-vue, .app-sidebar")||h.querySelector("[data-cy-files-sidebar], #app-sidebar-vue, .app-sidebar"))){w(60);return}});c.observe(document.body,{childList:!0,subtree:!0}),e=c},scheduleRefresh:w}},te=m+"_public_pad",ne="icon-filetype-etherpad-nextcloud-pad",ie='<svg xmlns="http://www.w3.org/2000/svg" xml:space="preserve" fill-rule="evenodd" stroke-linejoin="round" stroke-miterlimit="2" clip-rule="evenodd" viewBox="0 0 355 355"><path fill="#5ac395" d="M317 89v177c0 28-23 51-51 51H89c-28 0-51-23-51-51V89c0-28 23-51 51-51h177c28 0 51 23 51 51"/><path fill="#4aa57f" fill-rule="nonzero" d="M240 144q-4-3-8 0-6 6 0 9a36 36 0 0 1 0 52 6 6 0 0 0 3 12q3 0 5-3c19-18 19-50 0-70"/><path fill="#479a76" fill-rule="nonzero" d="M267 122q-4-5-9 0-4 4 0 10c25 26 25 68 0 93q-4 5 0 9 5 4 9 0c30-31 30-81 0-112"/><path fill="#fff" d="M192 130v2q-1 8-10 10H76q-9-2-11-10v-2q1-10 11-10h106q9 0 10 10m24 51v2q-1 11-12 11H78q-10 0-12-11v-2q2-10 12-11h126q12 1 12 11"/><path fill="#fff" d="M216 181v2q-1 11-12 11H78q-10 0-12-11v-2q2-10 12-11h126q12 1 12 11m-57 52v1q-1 11-10 11H76q-9 0-11-11v-1q1-10 11-11h73q9 1 10 11"/><path fill="#fff" d="M159 233v1q-1 11-10 11H76q-9 0-11-11v-1q1-10 11-11h73q9 1 10 11"/></svg>',re=98,mt=120,yt=500,oe=()=>{const e=xe();return e!==null&&e>=33},T=e=>(e&&e.message?String(e.message):"").toLowerCase().includes("duplicate"),wt=e=>{if(!e||typeof e!="object")return null;if(e.hasCreatePermission===!1)return!1;if(e.hasCreatePermission===!0)return!0;if(typeof e.canCreate=="boolean")return e.canCreate;const i=Ae(),n=[e.permissions,e.permission,e.attributes&&e.attributes.permissions,typeof e.get=="function"?e.get("permissions"):null];for(const r of n){const o=Number(r);if(Number.isFinite(o))return(o&i)===i}return null},bt=(...e)=>{for(const i of e){const n=wt(i);if(n!==null)return n}return!0},ht=()=>{const e=[window.OCP&&window.OCP.Files,window.OCA&&window.OCA.Files];for(const n of e)if(!(!n||typeof n!="object")){if(typeof n.addNewFileMenuEntry=="function")return{register:n.addNewFileMenuEntry.bind(n),unregister:typeof n.removeNewFileMenuEntry=="function"?n.removeNewFileMenuEntry.bind(n):null};if(typeof n.getNewFileMenu=="function"){let r=null;try{r=n.getNewFileMenu()}catch{r=null}if(r&&typeof r.registerEntry=="function")return{register:r.registerEntry.bind(r),unregister:typeof r.unregisterEntry=="function"?r.unregisterEntry.bind(r):null}}}const i=window._nc_newfilemenu;return i&&typeof i=="object"&&typeof i.registerEntry=="function"?{register:i.registerEntry.bind(i),unregister:typeof i.unregisterEntry=="function"?i.unregisterEntry.bind(i):null}:null},gt=({isFilesAppRoute:e,onCreatePublicPad:i})=>{let n=!1,r=!1,o=!1,a=0;const y=()=>{i()},s=(...w)=>bt(...w),p=()=>({id:te,displayName:t(m,"Public pad"),...oe()?{iconSvgInline:ie}:{iconClass:ne},order:re,enabled:s,handler:y}),l=()=>({id:te,displayName:t(m,"Public pad"),templateName:t(m,"Public pad"),iconClass:ne,...oe()?{iconSvgInline:ie}:{},fileType:"file",order:re,actionHandler:y}),u=async()=>{const w=ht();if(!w)return!1;const c=p();try{return w.register(c),n=!0,!0}catch(f){return T(f)?(n=!0,!0):!1}},b=()=>{if(r)return!0;const w=l(),c=()=>{const f=[window.OCA&&window.OCA.Files&&window.OCA.Files.NewFileMenu,window.OCA&&window.OCA.Files&&window.OCA.Files.newFileMenu];for(const d of f)if(!(!d||typeof d!="object")){if(typeof d.addMenuEntry=="function")try{return d.addMenuEntry(w),!0}catch(h){if(T(h))return!0}if(typeof d.registerEntry=="function")try{return d.registerEntry(p()),!0}catch(h){if(T(h))return!0}}return!1};try{return c()?(r=!0,!0):(!o&&window.OC&&window.OC.Plugins&&typeof window.OC.Plugins.register=="function"&&(window.OC.Plugins.register("OCA.Files.NewFileMenu",{attach(f){if(!(!f||typeof f.addMenuEntry!="function"))try{f.addMenuEntry(w),r=!0}catch(d){T(d)&&(r=!0)}}}),o=!0),r)}catch(f){return T(f)?(r=!0,!0):!1}};return()=>{if(!e()||n||r)return;const w=++a,c=async f=>{w!==a||n||r||await u()||b()||f<mt&&window.setTimeout(()=>{c(f+1)},yt)};c(0)}},vt=(e,i)=>{if(!(e instanceof HTMLAnchorElement)||!i)return null;const n=e.getAttribute("href")||"",r=Ne(n);if(r&&r.token===i&&R(r.path))return r.path;const o=Pe(n);return R(o)?o:null},Ct=({openPadInNativeViewer:e})=>{if(window.OCA&&window.OCA.EtherpadNextcloudClickInterceptorRegistered===!0)return;window.OCA&&(window.OCA.EtherpadNextcloudClickInterceptorRegistered=!0);const i=async n=>{if(!n||n.defaultPrevented)return;const r=_();if(!r||n.type==="click"&&(n.button!==0||n.metaKey||n.ctrlKey||n.shiftKey||n.altKey)||!(n.target instanceof Element))return;const o=n.target.closest("a[href]");if(!(o instanceof HTMLAnchorElement))return;const a=vt(o,r);if(a){if(n.preventDefault(),n.stopPropagation(),typeof n.stopImmediatePropagation=="function"&&n.stopImmediatePropagation(),M()){await e({path:a,fileId:null});return}window.location.assign(U(r,a))}};document.addEventListener("click",i,!0)},I="data-epnc-single-file-share";let ae=0;const be=e=>{if(!e)return null;const i='a[href*="/s/'+e+'/download"],a[href*="/index.php/s/'+e+'/download"]',n=document.querySelector(i);return n instanceof HTMLAnchorElement?n:null},Et=()=>{const e=document.getElementById("initial-state-files_sharing-view");if(!(e instanceof HTMLInputElement))return"";const i=String(e.value||"").trim();if(i==="")return"";try{const n=window.atob(i);if(n.startsWith('"')||n.startsWith("{")||n.startsWith("[")){const r=JSON.parse(n);return typeof r=="string"?r:""}return n}catch{return""}},he=e=>{if(!e)return!1;const i=be(e);if(!i)return!1;let n;try{n=new URL(i.getAttribute("href")||"",window.location.origin)}catch{return!1}const r=(n.pathname||"").match(/(?:\/index\.php)?\/s\/([^/]+)\/download\/?$/);return!r||r[1]!==e?!1:!n.searchParams.has("files")},xt=()=>{const e=document.body;if(!(e instanceof HTMLElement))return{retry:!0};const i=_();if(!i)return e.removeAttribute(I),{retry:!1};const n=Et();return n==="public-file-share"?(e.setAttribute(I,"1"),{retry:!1}):n==="public-share"?(e.removeAttribute(I),{retry:!1}):be(i)?he(i)?(e.setAttribute(I,"1"),{retry:!1}):(e.removeAttribute(I),{retry:!1}):(e.removeAttribute(I),{retry:!0})},At=()=>{const e=++ae,i=n=>{e===ae&&xt().retry&&n<30&&window.setTimeout(()=>i(n+1),180)};i(0)},se="EtherpadNextcloudRouteWatchInstalled",St=({ensurePublicPadMenuRegistration:e,openPadInNativeViewer:i,schedulePublicSingleShareUiStateRefresh:n,scheduleSidebarSyncPanelRefresh:r})=>{let o="";const a=async()=>{if(M())return;const l=q();if(!l)return;const u=new URLSearchParams(window.location.search||"");if(!(u.get("epNoRedirect")==="1"||u.get("details")==="1"||u.get("opendetails")==="true")&&!(window.location.pathname||"").includes("/apps/"+m+"/"))try{const b=await z(l);b&&b.is_pad&&b.viewer_url&&window.location.assign(b.viewer_url)}catch{}},y=async()=>{if(M()||!A())return;const l=q();if(!l)return;const u=new URLSearchParams(window.location.search||"");if(!(u.get("openfile")==="true"||u.get("opendetails")==="true"||u.get("details")==="1"))try{const b=await z(l);if(!b||!b.is_pad)return;const w=u.get("dir")||"/",c=L("/apps/files/files")+"?dir="+encodeURIComponent(w);window.location.href!==c&&window.location.assign(c)}catch{}},s=()=>{const l=_();if(!l)return;const u=new URLSearchParams(window.location.search||""),b=u.get("files")||"";if(M()){if(!R(b))return;const f=u.get("path")||"/",d=O(f,b);i({path:d,fileId:null});return}if(!R(b)){if((window.location.pathname||"").includes("/apps/"+m+"/public/"))return;he(l)&&window.location.assign(U(l,""));return}const w=u.get("path")||"/",c=O(w,b);window.location.assign(U(l,c))},p=()=>{const l=(window.location.pathname||"")+"?"+(window.location.search||"");l!==o&&(o=l,A()&&e(),n(),y(),a(),s(),r(0))};return{evaluateCurrentRoute:p,installRouteWatchers:()=>{if(window.OCA&&window.OCA[se]===!0)return;window.OCA&&(window.OCA[se]=!0);const l=()=>window.setTimeout(p,0);window.addEventListener("popstate",l),window.addEventListener("hashchange",l);const u=b=>{const w=window.history&&window.history[b];typeof w=="function"&&(window.history[b]=function(...c){const f=w.apply(this,c);return l(),f})};u("pushState"),u("replaceState")}}};(function(){let e=!1;const i=ft(),n=$e(),r=at({openPadInNativeViewer:n}),o=gt({isFilesAppRoute:A,onCreatePublicPad:r}),a=St({ensurePublicPadMenuRegistration:o,openPadInNativeViewer:n,schedulePublicSingleShareUiStateRefresh:At,scheduleSidebarSyncPanelRefresh:i.scheduleRefresh}),y=()=>{e||(e=!0,a.installRouteWatchers(),a.evaluateCurrentRoute(),Oe({openPadInNativeViewer:n}),Ct({openPadInNativeViewer:n}),i.installObserver(),i.scheduleRefresh(200))};document.readyState==="loading"?document.addEventListener("DOMContentLoaded",y,{once:!0}):y()})();
 //# sourceMappingURL=etherpad_nextcloud-files-main.mjs.map
