+++ js/etherpad_nextcloud-viewer-main.mjs
@@ -1,2 +1,2 @@
-import{V as S,M as L,t as m,c as y,A as u,o as v}from"./oc-compat-C8UpNj_V.chunk.mjs";import{b as I,e as _}from"./urls-C9SLQxSc.chunk.mjs";const b=d=>String(d||"").replace(/&/g,"&amp;").replace(/"/g,"&quot;").replace(/</g,"&lt;").replace(/>/g,"&gt;"),P=["http:","https:"],x=d=>{try{const f=new URL(String(d||""));return P.includes(f.protocol)&&f.username===""&&f.password===""}catch{return!1}},H="default-src 'none'; frame-src http: https:; style-src 'unsafe-inline'",O=d=>{const f=x(d)?b(d):"";return'<!doctype html><html><head><meta charset="utf-8"><meta http-equiv="Content-Security-Policy" content="'+b(H)+'"><style>html,body,iframe{width:100%;height:100%;margin:0;border:0;overflow:hidden}iframe{display:block}</style></head><body><iframe src="'+f+'" title="Etherpad"></iframe></body></html>'};(function(){let d=0;const f={name:"EtherpadNextcloudViewer",props:{filename:{type:String,required:!1,default:""},basename:{type:String,required:!1,default:""},source:{type:String,required:!1,default:""},fileid:{type:[String,Number],required:!1,default:null},fileId:{type:[String,Number],required:!1,default:null},fileInfo:{type:Object,required:!1,default:null}},data(){return{iframeSrc:"",isLoading:!0,loadError:"",externalOpenUrl:"",externalOpenMessage:"",snapshotMode:"",snapshot:{text:"",html:""},resolveGeneration:0,syncUrl:"",syncIntervalMs:12e4,syncInFlight:!1,syncTimerId:null,visibilityHandler:null,pageHideHandler:null}},computed:{sourcePath(){const e=typeof this.source=="string"?this.source.trim():"";return e&&_(e)||""},filePath(){const e=t=>String(t||"").trim().replace(/\s+\.pad$/i,".pad"),s=t=>{const a=String(t||"").trim();return!a||a==="/"?"/":a.startsWith("/")?a:"/"+a},n=(t,a)=>{if(!a)return"";if(a.startsWith("/"))return a;const o=s(t);return o==="/"?"/"+a:o+"/"+a},r=t=>typeof t=="string"&&t.toLowerCase().endsWith(".pad");if(r(this.sourcePath))return this.sourcePath;const i=this.fileInfo&&typeof this.fileInfo=="object"?this.fileInfo:null,c=i&&typeof i.path=="string"?e(i.path):"";if(r(c))return c.startsWith("/")?c:"/"+c;const h=e(this.filename||this.basename||i&&(i.name||i.basename)||"");if(!h)return"";const p=i&&typeof i.dirname=="string"?i.dirname:"";if(p){const t=n(p,h);if(r(t))return t}const g=new URLSearchParams(window.location.search||"").get("dir")||"/",l=n(g,h);return r(l)?l:"/"+h},resolvedFileId(){const e=[this.fileid,this.fileId,this.fileInfo&&(this.fileInfo.fileid||this.fileInfo.fileId||this.fileInfo.id)];for(const r of e){const i=Number(r);if(Number.isFinite(i)&&i>0)return i}if(new URLSearchParams(window.location.search||"").get("openfile")!=="true")return null;const s=(window.location.pathname||"").match(/\/apps\/files\/files\/(\d+)\/?$/);if(!s)return null;const n=Number(s[1]);return Number.isFinite(n)&&n>0?n:null}},watch:{filePath:{immediate:!0,handler(){this.resolveOpenUrl()}},resolvedFileId(){this.resolveOpenUrl()}},methods:{async fetchOpenPayload(e,s={}){const n=Object.assign({Accept:"application/json"},s.headers||{}),r=await fetch(e,Object.assign({method:"GET",credentials:"same-origin",headers:n},s)),i=await r.json().catch(()=>({}));if(!r.ok)throw new Error(i&&i.message||"Pad open failed.");if(!i||i.is_readonly_snapshot!==!0&&(typeof i.url!="string"||i.url.trim()===""))throw new Error("Pad open API did not return a valid URL.");return i},isMissingFrontmatterError(e){return e instanceof Error?String(e.message||"").includes("Missing YAML frontmatter"):!1},async initializeMissingFrontmatter(){const e={Accept:"application/json",requesttoken:v()};if(this.resolvedFileId!==null){const c=y("/apps/"+u+"/api/v1/pads/initialize-by-id/"+encodeURIComponent(String(this.resolvedFileId))),h=await fetch(c,{method:"POST",credentials:"same-origin",headers:e}),p=await h.json().catch(()=>({}));if(!h.ok)throw new Error(p&&p.message||"Pad initialization failed.");return p}if(!this.filePath)throw new Error("Pad initialization failed: missing file path.");const s=new URLSearchParams;s.set("file",this.filePath);const n=y("/apps/"+u+"/api/v1/pads/initialize"),r=await fetch(n,{method:"POST",credentials:"same-origin",headers:Object.assign({},e,{"Content-Type":"application/x-www-form-urlencoded;charset=UTF-8"}),body:s.toString()}),i=await r.json().catch(()=>({}));if(!r.ok)throw new Error(i&&i.message||"Pad initialization failed.");return i},markLoaded(){this.$emit("update:loaded",!0)},stopSyncLoop(){this.syncTimerId!==null&&(window.clearInterval(this.syncTimerId),this.syncTimerId=null)},startSyncLoop(){!this.syncUrl||this.syncTimerId!==null||(this.syncTimerId=window.setInterval(()=>{document.visibilityState==="visible"&&this.runSync(!1,!1)},this.syncIntervalMs))},async runSync(e,s){if(this.syncUrl&&!(this.syncInFlight&&!e)){this.syncInFlight=!0;try{const n=e?this.syncUrl+(this.syncUrl.includes("?")?"&":"?")+"force=1":this.syncUrl;await fetch(n,{method:"POST",credentials:"same-origin",headers:{Accept:"application/json",requesttoken:v()},keepalive:!!s})}finally{this.syncInFlight=!1}}},installSyncLifecycleHandlers(){this.visibilityHandler||this.pageHideHandler||(this.visibilityHandler=()=>{if(document.visibilityState==="hidden"){this.runSync(!0,!0),this.stopSyncLoop();return}this.startSyncLoop()},this.pageHideHandler=()=>{this.runSync(!0,!0),this.stopSyncLoop()},document.addEventListener("visibilitychange",this.visibilityHandler),window.addEventListener("pagehide",this.pageHideHandler))},removeSyncLifecycleHandlers(){this.visibilityHandler&&(document.removeEventListener("visibilitychange",this.visibilityHandler),this.visibilityHandler=null),this.pageHideHandler&&(window.removeEventListener("pagehide",this.pageHideHandler),this.pageHideHandler=null)},async resolveOpenUrl(){const e=++this.resolveGeneration,s=()=>e===this.resolveGeneration;if(this.isLoading=!0,this.loadError="",this.iframeSrc="",this.externalOpenUrl="",this.externalOpenMessage="",this.snapshotMode="",this.snapshot={text:"",html:""},this.syncUrl="",this.syncInFlight=!1,this.stopSyncLoop(),!this.filePath){if(!s())return;this.loadError="No .pad file selected.",this.isLoading=!1;return}const n=I(),r=y("/apps/"+u+"/api/v1/pads/open"),i=(()=>{if(!n)return"";const l=new URL(y("/apps/"+u+"/api/v1/public/open/"+encodeURIComponent(n)),window.location.origin);return l.searchParams.set("file",this.filePath),l.toString()})(),c=this.resolvedFileId!==null?y("/apps/"+u+"/api/v1/pads/open-by-id"):"",h=new URLSearchParams;h.set("file",this.filePath);const p=new URLSearchParams;this.resolvedFileId!==null&&p.set("fileId",String(this.resolvedFileId));const g={"Content-Type":"application/x-www-form-urlencoded;charset=UTF-8",requesttoken:v()};try{const l=async()=>{let o=null;if(i)o=await this.fetchOpenPayload(i);else if(c)try{o=await this.fetchOpenPayload(c,{method:"POST",headers:g,body:p.toString()})}catch{}return o||(o=await this.fetchOpenPayload(r,{method:"POST",headers:g,body:h.toString()})),o};let t;try{if(t=await l(),!s())return}catch(o){if(!this.isMissingFrontmatterError(o))throw o;if(await this.initializeMissingFrontmatter(),!s()||(t=await l(),!s()))return}this.syncUrl=t&&typeof t.sync_url=="string"?t.sync_url:"";const a=Number(t&&t.sync_interval_seconds);if(this.syncIntervalMs=Number.isFinite(a)&&a>0?Math.max(5e3,Math.min(36e5,a*1e3)):12e4,this.installSyncLifecycleHandlers(),this.syncUrl&&this.startSyncLoop(),t&&t.is_readonly_snapshot===!0){this.snapshotMode="readonly",this.snapshot={text:typeof t.snapshot_text=="string"?t.snapshot_text:"",html:typeof t.snapshot_html=="string"?t.snapshot_html:""},this.markLoaded();return}if(t&&t.is_external===!0&&typeof t.url=="string"&&t.url.trim()!==""){const o=t.url.trim();this.externalOpenUrl=o,this.externalOpenMessage=m("Read-only snapshot from the .pad file."),this.snapshotMode="external",this.snapshot={text:t&&typeof t.snapshot_text=="string"?t.snapshot_text:"",html:t&&typeof t.snapshot_html=="string"?t.snapshot_html:""},this.markLoaded();return}this.iframeSrc=t.url,this.markLoaded()}catch(l){if(!s())return;this.loadError=l instanceof Error?l.message:"Could not load pad.",this.markLoaded()}finally{if(!s())return;this.isLoading=!1}},renderSnapshotView(e,s){const n=String(s.html||""),r=String(s.text||""),i=Array.isArray(s.actions)?s.actions:[];return e("div",{class:"epnc-native-snapshot"},[e("div",{class:"epnc-native-snapshot__inner"},[e("div",{class:"epnc-native-snapshot__header"},[e("div",{class:"epnc-native-snapshot__heading"},[e("div",{class:"epnc-native-snapshot__title"},s.title),e("div",{class:"epnc-native-snapshot__message"},s.message)]),i.length>0?e("div",{class:"epnc-native-snapshot__actions"},i):null]),n.trim()!==""?e("div",{class:"epnc-native-snapshot__text epnc-native-snapshot__text--html",domProps:{innerHTML:n}}):e("pre",{class:"epnc-native-snapshot__text"},r.trim()!==""?r:s.emptyMessage)])])}},beforeDestroy(){this.resolveGeneration+=1,this.runSync(!0,!0),this.stopSyncLoop(),this.removeSyncLifecycleHandlers()},beforeUnmount(){this.resolveGeneration+=1,this.runSync(!0,!0),this.stopSyncLoop(),this.removeSyncLifecycleHandlers()},render(e){return this.loadError?e("div",{class:"epnc-native-status epnc-native-status--error"},[e("div",{class:"epnc-native-error-card"},[e("div",{class:"epnc-native-error-title"},"Unable to open pad"),e("div",{class:"epnc-native-error-message"},this.loadError)])]):this.snapshotMode==="external"?this.renderSnapshotView(e,{title:m("Pad from another server"),message:this.externalOpenMessage,actions:[e("a",{class:"button primary",attrs:{href:this.externalOpenUrl,target:"_blank",rel:"noopener noreferrer"}},m("Open original pad"))],html:this.snapshot.html,text:this.snapshot.text,emptyMessage:m("No synced snapshot is stored in this .pad file yet.")}):this.snapshotMode==="readonly"?this.renderSnapshotView(e,{title:m("Read-only snapshot"),message:m("Read-only snapshot from the .pad file."),html:this.snapshot.html,text:this.snapshot.text,emptyMessage:m("No synced snapshot is stored in this .pad file yet.")}):this.isLoading||!this.iframeSrc?e("div",{class:"epnc-native-status"},"Loading pad..."):e("div",{class:"epnc-native-shell"},[e("iframe",{attrs:{srcdoc:O(this.iframeSrc),title:"Etherpad"},on:{load:()=>this.markLoaded(),error:()=>this.markLoaded()},class:"epnc-native-iframe"})])}},w=()=>{if(d+=1,!(window.OCA&&window.OCA.Viewer&&typeof window.OCA.Viewer.registerHandler=="function")){d<20&&window.setTimeout(w,500);return}Array.isArray(window.OCA.Viewer.availableHandlers)&&window.OCA.Viewer.availableHandlers.some(e=>e&&e.id===S)||window.OCA.Viewer.registerHandler({id:S,mimes:[L],component:f})};w()})();
+import{V as S,M as L,t as m,c as y,A as u,o as v}from"./oc-compat-C8UpNj_V.chunk.mjs";import{a as I,e as _}from"./urls-BqezDGeR.chunk.mjs";const b=d=>String(d||"").replace(/&/g,"&amp;").replace(/"/g,"&quot;").replace(/</g,"&lt;").replace(/>/g,"&gt;"),P=["http:","https:"],x=d=>{try{const f=new URL(String(d||""));return P.includes(f.protocol)&&f.username===""&&f.password===""}catch{return!1}},H="default-src 'none'; frame-src http: https:; style-src 'unsafe-inline'",O=d=>{const f=x(d)?b(d):"";return'<!doctype html><html><head><meta charset="utf-8"><meta http-equiv="Content-Security-Policy" content="'+b(H)+'"><style>html,body,iframe{width:100%;height:100%;margin:0;border:0;overflow:hidden}iframe{display:block}</style></head><body><iframe src="'+f+'" title="Etherpad"></iframe></body></html>'};(function(){let d=0;const f={name:"EtherpadNextcloudViewer",props:{filename:{type:String,required:!1,default:""},basename:{type:String,required:!1,default:""},source:{type:String,required:!1,default:""},fileid:{type:[String,Number],required:!1,default:null},fileId:{type:[String,Number],required:!1,default:null},fileInfo:{type:Object,required:!1,default:null}},data(){return{iframeSrc:"",isLoading:!0,loadError:"",externalOpenUrl:"",externalOpenMessage:"",snapshotMode:"",snapshot:{text:"",html:""},resolveGeneration:0,syncUrl:"",syncIntervalMs:12e4,syncInFlight:!1,syncTimerId:null,visibilityHandler:null,pageHideHandler:null}},computed:{sourcePath(){const e=typeof this.source=="string"?this.source.trim():"";return e&&_(e)||""},filePath(){const e=t=>String(t||"").trim().replace(/\s+\.pad$/i,".pad"),s=t=>{const a=String(t||"").trim();return!a||a==="/"?"/":a.startsWith("/")?a:"/"+a},n=(t,a)=>{if(!a)return"";if(a.startsWith("/"))return a;const o=s(t);return o==="/"?"/"+a:o+"/"+a},r=t=>typeof t=="string"&&t.toLowerCase().endsWith(".pad");if(r(this.sourcePath))return this.sourcePath;const i=this.fileInfo&&typeof this.fileInfo=="object"?this.fileInfo:null,c=i&&typeof i.path=="string"?e(i.path):"";if(r(c))return c.startsWith("/")?c:"/"+c;const h=e(this.filename||this.basename||i&&(i.name||i.basename)||"");if(!h)return"";const p=i&&typeof i.dirname=="string"?i.dirname:"";if(p){const t=n(p,h);if(r(t))return t}const g=new URLSearchParams(window.location.search||"").get("dir")||"/",l=n(g,h);return r(l)?l:"/"+h},resolvedFileId(){const e=[this.fileid,this.fileId,this.fileInfo&&(this.fileInfo.fileid||this.fileInfo.fileId||this.fileInfo.id)];for(const r of e){const i=Number(r);if(Number.isFinite(i)&&i>0)return i}if(new URLSearchParams(window.location.search||"").get("openfile")!=="true")return null;const s=(window.location.pathname||"").match(/\/apps\/files\/files\/(\d+)\/?$/);if(!s)return null;const n=Number(s[1]);return Number.isFinite(n)&&n>0?n:null}},watch:{filePath:{immediate:!0,handler(){this.resolveOpenUrl()}},resolvedFileId(){this.resolveOpenUrl()}},methods:{async fetchOpenPayload(e,s={}){const n=Object.assign({Accept:"application/json"},s.headers||{}),r=await fetch(e,Object.assign({method:"GET",credentials:"same-origin",headers:n},s)),i=await r.json().catch(()=>({}));if(!r.ok)throw new Error(i&&i.message||"Pad open failed.");if(!i||i.is_readonly_snapshot!==!0&&(typeof i.url!="string"||i.url.trim()===""))throw new Error("Pad open API did not return a valid URL.");return i},isMissingFrontmatterError(e){return e instanceof Error?String(e.message||"").includes("Missing YAML frontmatter"):!1},async initializeMissingFrontmatter(){const e={Accept:"application/json",requesttoken:v()};if(this.resolvedFileId!==null){const c=y("/apps/"+u+"/api/v1/pads/initialize-by-id/"+encodeURIComponent(String(this.resolvedFileId))),h=await fetch(c,{method:"POST",credentials:"same-origin",headers:e}),p=await h.json().catch(()=>({}));if(!h.ok)throw new Error(p&&p.message||"Pad initialization failed.");return p}if(!this.filePath)throw new Error("Pad initialization failed: missing file path.");const s=new URLSearchParams;s.set("file",this.filePath);const n=y("/apps/"+u+"/api/v1/pads/initialize"),r=await fetch(n,{method:"POST",credentials:"same-origin",headers:Object.assign({},e,{"Content-Type":"application/x-www-form-urlencoded;charset=UTF-8"}),body:s.toString()}),i=await r.json().catch(()=>({}));if(!r.ok)throw new Error(i&&i.message||"Pad initialization failed.");return i},markLoaded(){this.$emit("update:loaded",!0)},stopSyncLoop(){this.syncTimerId!==null&&(window.clearInterval(this.syncTimerId),this.syncTimerId=null)},startSyncLoop(){!this.syncUrl||this.syncTimerId!==null||(this.syncTimerId=window.setInterval(()=>{document.visibilityState==="visible"&&this.runSync(!1,!1)},this.syncIntervalMs))},async runSync(e,s){if(this.syncUrl&&!(this.syncInFlight&&!e)){this.syncInFlight=!0;try{const n=e?this.syncUrl+(this.syncUrl.includes("?")?"&":"?")+"force=1":this.syncUrl;await fetch(n,{method:"POST",credentials:"same-origin",headers:{Accept:"application/json",requesttoken:v()},keepalive:!!s})}finally{this.syncInFlight=!1}}},installSyncLifecycleHandlers(){this.visibilityHandler||this.pageHideHandler||(this.visibilityHandler=()=>{if(document.visibilityState==="hidden"){this.runSync(!0,!0),this.stopSyncLoop();return}this.startSyncLoop()},this.pageHideHandler=()=>{this.runSync(!0,!0),this.stopSyncLoop()},document.addEventListener("visibilitychange",this.visibilityHandler),window.addEventListener("pagehide",this.pageHideHandler))},removeSyncLifecycleHandlers(){this.visibilityHandler&&(document.removeEventListener("visibilitychange",this.visibilityHandler),this.visibilityHandler=null),this.pageHideHandler&&(window.removeEventListener("pagehide",this.pageHideHandler),this.pageHideHandler=null)},async resolveOpenUrl(){const e=++this.resolveGeneration,s=()=>e===this.resolveGeneration;if(this.isLoading=!0,this.loadError="",this.iframeSrc="",this.externalOpenUrl="",this.externalOpenMessage="",this.snapshotMode="",this.snapshot={text:"",html:""},this.syncUrl="",this.syncInFlight=!1,this.stopSyncLoop(),!this.filePath){if(!s())return;this.loadError="No .pad file selected.",this.isLoading=!1;return}const n=I(),r=y("/apps/"+u+"/api/v1/pads/open"),i=(()=>{if(!n)return"";const l=new URL(y("/apps/"+u+"/api/v1/public/open/"+encodeURIComponent(n)),window.location.origin);return l.searchParams.set("file",this.filePath),l.toString()})(),c=this.resolvedFileId!==null?y("/apps/"+u+"/api/v1/pads/open-by-id"):"",h=new URLSearchParams;h.set("file",this.filePath);const p=new URLSearchParams;this.resolvedFileId!==null&&p.set("fileId",String(this.resolvedFileId));const g={"Content-Type":"application/x-www-form-urlencoded;charset=UTF-8",requesttoken:v()};try{const l=async()=>{let o=null;if(i)o=await this.fetchOpenPayload(i);else if(c)try{o=await this.fetchOpenPayload(c,{method:"POST",headers:g,body:p.toString()})}catch{}return o||(o=await this.fetchOpenPayload(r,{method:"POST",headers:g,body:h.toString()})),o};let t;try{if(t=await l(),!s())return}catch(o){if(!this.isMissingFrontmatterError(o))throw o;if(await this.initializeMissingFrontmatter(),!s()||(t=await l(),!s()))return}this.syncUrl=t&&typeof t.sync_url=="string"?t.sync_url:"";const a=Number(t&&t.sync_interval_seconds);if(this.syncIntervalMs=Number.isFinite(a)&&a>0?Math.max(5e3,Math.min(36e5,a*1e3)):12e4,this.installSyncLifecycleHandlers(),this.syncUrl&&this.startSyncLoop(),t&&t.is_readonly_snapshot===!0){this.snapshotMode="readonly",this.snapshot={text:typeof t.snapshot_text=="string"?t.snapshot_text:"",html:typeof t.snapshot_html=="string"?t.snapshot_html:""},this.markLoaded();return}if(t&&t.is_external===!0&&typeof t.url=="string"&&t.url.trim()!==""){const o=t.url.trim();this.externalOpenUrl=o,this.externalOpenMessage=m("Read-only snapshot from the .pad file."),this.snapshotMode="external",this.snapshot={text:t&&typeof t.snapshot_text=="string"?t.snapshot_text:"",html:t&&typeof t.snapshot_html=="string"?t.snapshot_html:""},this.markLoaded();return}this.iframeSrc=t.url,this.markLoaded()}catch(l){if(!s())return;this.loadError=l instanceof Error?l.message:"Could not load pad.",this.markLoaded()}finally{if(!s())return;this.isLoading=!1}},renderSnapshotView(e,s){const n=String(s.html||""),r=String(s.text||""),i=Array.isArray(s.actions)?s.actions:[];return e("div",{class:"epnc-native-snapshot"},[e("div",{class:"epnc-native-snapshot__inner"},[e("div",{class:"epnc-native-snapshot__header"},[e("div",{class:"epnc-native-snapshot__heading"},[e("div",{class:"epnc-native-snapshot__title"},s.title),e("div",{class:"epnc-native-snapshot__message"},s.message)]),i.length>0?e("div",{class:"epnc-native-snapshot__actions"},i):null]),n.trim()!==""?e("div",{class:"epnc-native-snapshot__text epnc-native-snapshot__text--html",domProps:{innerHTML:n}}):e("pre",{class:"epnc-native-snapshot__text"},r.trim()!==""?r:s.emptyMessage)])])}},beforeDestroy(){this.resolveGeneration+=1,this.runSync(!0,!0),this.stopSyncLoop(),this.removeSyncLifecycleHandlers()},beforeUnmount(){this.resolveGeneration+=1,this.runSync(!0,!0),this.stopSyncLoop(),this.removeSyncLifecycleHandlers()},render(e){return this.loadError?e("div",{class:"epnc-native-status epnc-native-status--error"},[e("div",{class:"epnc-native-error-card"},[e("div",{class:"epnc-native-error-title"},"Unable to open pad"),e("div",{class:"epnc-native-error-message"},this.loadError)])]):this.snapshotMode==="external"?this.renderSnapshotView(e,{title:m("Pad from another server"),message:this.externalOpenMessage,actions:[e("a",{class:"button primary",attrs:{href:this.externalOpenUrl,target:"_blank",rel:"noopener noreferrer"}},m("Open original pad"))],html:this.snapshot.html,text:this.snapshot.text,emptyMessage:m("No synced snapshot is stored in this .pad file yet.")}):this.snapshotMode==="readonly"?this.renderSnapshotView(e,{title:m("Read-only snapshot"),message:m("Read-only snapshot from the .pad file."),html:this.snapshot.html,text:this.snapshot.text,emptyMessage:m("No synced snapshot is stored in this .pad file yet.")}):this.isLoading||!this.iframeSrc?e("div",{class:"epnc-native-status"},"Loading pad..."):e("div",{class:"epnc-native-shell"},[e("iframe",{attrs:{srcdoc:O(this.iframeSrc),title:"Etherpad"},on:{load:()=>this.markLoaded(),error:()=>this.markLoaded()},class:"epnc-native-iframe"})])}},w=()=>{if(d+=1,!(window.OCA&&window.OCA.Viewer&&typeof window.OCA.Viewer.registerHandler=="function")){d<20&&window.setTimeout(w,500);return}Array.isArray(window.OCA.Viewer.availableHandlers)&&window.OCA.Viewer.availableHandlers.some(e=>e&&e.id===S)||window.OCA.Viewer.registerHandler({id:S,mimes:[L],component:f})};w()})();
 //# sourceMappingURL=etherpad_nextcloud-viewer-main.mjs.map
